% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAfforestationMask.R
\name{calcAfforestationMask}
\alias{calcAfforestationMask}
\title{calcAfforestationMask}
\usage{
calcAfforestationMask(subtype, cells = "lpjcell")
}
\arguments{
\item{subtype}{afforestation mask sub type}

\item{cells}{"magpiecell" or "lpjcell"}
}
\value{
magpie object in cellular resolution
}
\description{
Afforestation mask for where afforestation possible
}
\examples{
\dontrun{
calcOutput("AfforestationMask", subtype = "noboreal", aggregate = FALSE)
}

}
\author{
David Chen, Florian Humpenoeder
}
