% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAreaEquippedForIrrigation.R
\name{calcAreaEquippedForIrrigation}
\alias{calcAreaEquippedForIrrigation}
\title{calcAreaEquippedForIrrigation}
\usage{
calcAreaEquippedForIrrigation(
  cellular = FALSE,
  cells = "lpjcell",
  selectyears = "past"
)
}
\arguments{
\item{cellular}{if TRUE: 0.5 degree resolution returned}

\item{cells}{number of cells to be returned: magpiecell (59199), lpjcell (67420)}

\item{selectyears}{default on "past"}
}
\value{
List of magpie objects with results on country/cellular level,
        weight on country level, unit and description.
}
\description{
Calculates the area equipped for irrigation based on LU2v2 or
             Mehta data sets.
             For LUH2v2, it assumes, that all cropland irrigated in the last
             20 years at least once is equipped for irrigation.
             Mehta et al. (2022) directly report Global Area Equipped for
             Irrigation for the years 1900-2015
}
\examples{
\dontrun{
calcOutput("AreaEquippedForIrrigation", source = "LUH2v2", cellular = TRUE, aggregate = FALSE)
}
}
\seealso{
[calcLanduseInitialisation()]
}
\author{
Benjamin Leon Bodirsky, Kristine Karstens, Felicitas Beier
}
