% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBinnedLsuDensity.R
\name{calcBinnedLsuDensity}
\alias{calcBinnedLsuDensity}
\title{calcRangeSoilCarbonHist}
\usage{
calcBinnedLsuDensity(
  breaks = c(seq(0, 2, 0.1), 2.25, 2.5),
  labels = c(0, 0.2, 0.2, 0.4, 0.4, 0.6, 0.6, 0.8, 0.8, 1, 1, 1.2, 1.2, 1.4, 1.4, 1.6,
    1.6, 1.8, 1.8, 2, 2, 2.5),
  years = 1995
)
}
\arguments{
\item{breaks}{Binning breaks}

\item{labels}{Binning labels}

\item{years}{years where data should binned}
}
\value{
Magpie object with lsu per cell.
}
\description{
calculates soil carbon for rangelands
}
\examples{
\dontrun{
calcOutput("BinnedLsuDensity ", breaks, labels, years)
}
}
\author{
Marcos Alves
}
