% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBphEffect.R
\name{calcBphEffect}
\alias{calcBphEffect}
\title{calcBphEffect}
\usage{
calcBphEffect(cells = "lpjcell")
}
\arguments{
\item{cells}{lpjcell for 67420 cells or magpiecell for 59199 cells}
}
\value{
magpie object in cellular resolution
}
\description{
Biogeophysical temperature change of afforestation (degree C).
             File is based on observation datasets of Bright et al. 2017
             and Duveiller et al. 2018
}
\examples{
\dontrun{
calcOutput("BphEffect", aggregate = FALSE)
}

}
\author{
Michael Windisch, Felicitas Beier
}
