% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBphMask.R
\name{calcBphMask}
\alias{calcBphMask}
\title{calcBphMask}
\usage{
calcBphMask(cells = "lpjcell")
}
\arguments{
\item{cells}{lpjcell for 67420 cells or magpiecell for 59199 cells}
}
\value{
magpie object in cellular resolution
}
\description{
Mask of Datapoints of biogeophysical temperature change of afforestation (degree C)
             to be used as weight.
             File is based on observation datasets of Bright et al. 2017 and Duveiller et al. 2018
}
\examples{
\dontrun{
calcOutput("BphMask", aggregate = FALSE)
}

}
\author{
Michael Windisch, Felicitas Beier
}
