% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbon.R
\name{calcCarbon}
\alias{calcCarbon}
\title{calcCarbon}
\usage{
calcCarbon(lpjml = "lpjml5.9.5-m1", climatetype = "MRI-ESM2-0:ssp370")
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different GCM climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
Extract the carbon densities used in MAgPIE from a set of LPJmL runs
}
\examples{
\dontrun{
calcOutput("Carbon", aggregate = FALSE)
}
}
\author{
Kristine Karstens, Patrick v. Jeetze, Michael Crawford
}
