% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonFromFlowHarm.R
\name{calcCarbonFromFlowHarm}
\alias{calcCarbonFromFlowHarm}
\title{calcCarbonFromFlowHarm}
\usage{
calcCarbonFromFlowHarm(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_44ac93de", crop =
    "ggcmi_phase3_nchecks_9ca735cb"),
  climatetype = "GSWP3-W5E5:historical"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different GCM climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates carbon sink per year and uses them for smoothing and harmonization
             and calculates carbon from it for MAgPIE again (to avoid jumps in carbon sink estimates)
}
\examples{
\dontrun{
calcOutput("calcCarbonFromFlowHarm", aggregate = FALSE)
}

}
\author{
Kristine Karstens
}
