% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbonSink.R
\name{calcCarbonSink}
\alias{calcCarbonSink}
\title{calcCarbonSink}
\usage{
calcCarbonSink(
  version = "LPJmL4_for_MAgPIE_44ac93de",
  climatetype = "GSWP3-W5E5:historical",
  stage = "harmonized2020"
)
}
\arguments{
\item{version}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different GCM climate scenarios}

\item{stage}{Degree of processing: raw, smoothed, harmonized, harmonized2020}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates carbon sink per year and smooths the data
             and calculates carbon from it for MAgPIE again (to avoid jumps in carbon sink estimates)
}
\examples{
\dontrun{ calcOutput("Carbon", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
