% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCellArea.R
\name{calcCellArea}
\alias{calcCellArea}
\title{calcCellArea}
\usage{
calcCellArea(cells = "magpiecell")
}
\arguments{
\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
magpie object in cellular resolution
}
\description{
total cell area
}
\examples{
\dontrun{
calcOutput("CellArea", aggregate = FALSE)
}

}
\author{
Jan Philipp Dietrich
}
