% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCellCountryFraction.R
\name{calcCellCountryFraction}
\alias{calcCellCountryFraction}
\title{calcCellCountryFraction}
\usage{
calcCellCountryFraction()
}
\value{
Clustered MAgPIE object on requested resolution
}
\description{
cell fraction belonging to a country based on LanduseInitialisation
}
\examples{
\dontrun{
calcOutput("calcCellCountryFraction", aggregate = FALSE)
}

}
\author{
Florian Humpenoeder
}
