% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCO2Atmosphere.R
\name{calcCO2Atmosphere}
\alias{calcCO2Atmosphere}
\title{calcCO2Atmosphere}
\usage{
calcCO2Atmosphere(rcp = "rcp8p5", co2_evolution = "rising")
}
\arguments{
\item{rcp}{specify the RCP (rcp85, rcp60, rcp45 or rcp26)}

\item{co2_evolution}{Define `rising` for rising CO2 according to the climate scenario selected or
`static` for stable CO2 at the last past time step level.}
}
\value{
magpie object in cellular resolution
}
\description{
Disaggregate CO2 global atmospheric concentration to cellular level
}
\examples{
\dontrun{
calcOutput("CO2Atmosphere", aggregate = FALSE, rcp = "rcp8p5", co2_evolution = "rising")
}

}
\author{
Marcos Alves, Kristine Karstens
}
