% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCO2Atmosphere_new.R
\name{calcCO2Atmosphere_new}
\alias{calcCO2Atmosphere_new}
\title{calcCO2Atmosphere_new}
\usage{
calcCO2Atmosphere_new(subtype = "ISIMIP3b:ssp126", co2_evolution = "rising")
}
\arguments{
\item{subtype}{specify the version and scenario eg. "ISIMIP3b:ssp126"}

\item{co2_evolution}{Define `rising` for rising CO2 according to the climate scenario selected or
`static` for stable CO2 at the last past time step level.}
}
\value{
magpie object in cellular resolution
}
\description{
Disaggregate CO2 global atmospheric concentration to cellular level
}
\examples{
\dontrun{
calcOutput("CO2Atmosphere_new", aggregate = FALSE, subtype, co2_evolution)
}

}
\author{
Marcos Alves, Kristine Karstens
}
