% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCollectEnvironmentData_new.R
\name{calcCollectEnvironmentData_new}
\alias{calcCollectEnvironmentData_new}
\title{calcCollectEnvironmentData_new}
\usage{
calcCollectEnvironmentData_new(
  subtype = "ISIMIP3b:IPSL-CM6A-LR:ssp126:1965-2100",
  sar = 20,
  sel_feat = c("tas", "pr", "lwnet", "rsds", "CO2", "Ks", "Sf", "w_pwp", "w_fc",
    "w_sat", "hsg", "wet")
)
}
\arguments{
\item{subtype}{Switch between different climate scenarios (default: "CRU_4") eg. "ISIMIP3b:IPSL-CM6A-LR:ssp126:1965-2100"}

\item{sar}{Average range for smoothing annual variations}

\item{sel_feat}{features names to be included in the output file}
}
\value{
magpie object in cellular resolution
}
\description{
Calculate climate, CO2 and soil environmental conditions on cellular level
}
\examples{
\dontrun{
calcOutput("CollectEnvironmentData_new", subtype, sar = 20, sel_feat = "temp")
}

}
\author{
Marcos Alves
}
