% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCollectSoilCarbonLSU.R
\name{calcCollectSoilCarbonLSU}
\alias{calcCollectSoilCarbonLSU}
\title{calcCollectSoilCarbonLSU}
\usage{
calcCollectSoilCarbonLSU(
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  lpjml = "LPJmL_cgrazing",
  climatetype = "HadGEM2_ES:rcp8p5:co2",
  sar = 20
)
}
\arguments{
\item{lsu_levels}{Livestock unit levels in the source folder}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{sar}{Average range for smoothing annual variations}
}
\value{
magpie object in cellular resolution
}
\description{
Calculate soil carbon stocks for different LSU and climate conditions
}
\examples{
\dontrun{
calcOutput("CollectSoilCarbonLSU ", lsu_levels = c(seq(0, 2, 0.2), 2.5))
}

}
\author{
Marcos Alves
}
