% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCollectSoilCarbonPastr.R
\name{calcCollectSoilCarbonPastr}
\alias{calcCollectSoilCarbonPastr}
\title{calcCollectSoilCarbonPastr}
\usage{
calcCollectSoilCarbonPastr(
  past_mngmt = "me2",
  lpjml = "lpjml5p2_pasture",
  climatetype = "IPSL_CM6A_LR",
  scenario = "ssp126_co2_limN",
  sar = 20
)
}
\arguments{
\item{past_mngmt}{pasture areas management option}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{scenario}{scenario specifications (eg. ssp126_co2_limN)}

\item{sar}{Average range for smoothing annual variations}
}
\value{
magpie object in cellular resolution
}
\description{
calculates soil carbon content for pasture areas
}
\examples{
\dontrun{
calcOutput("CollectSoilCarbonPastr", past_mngmt = "me2")
}


}
\author{
Marcos Alves
}
