% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCommittedAgWaterUse.R
\name{calcCommittedAgWaterUse}
\alias{calcCommittedAgWaterUse}
\title{calcCommittedAgWaterUse}
\usage{
calcCommittedAgWaterUse(
  version = "LPJmL5",
  iniyear = 1995,
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "spline",
  dof = 4,
  averaging_range = NULL
)
}
\arguments{
\item{version}{Switch between LPJmL4 and LPJmL5}

\item{iniyear}{Year of initialization}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{Time smoothing: average or spline (default)}

\item{dof}{Only specify if time=="spline": degrees of freedom needed for spline}

\item{averaging_range}{Only specify if time=="average": number of time steps to average}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates committed agricultural water uses that are used in the river routing algorithm for distributing available water across the basin
}
\examples{
\dontrun{ calcOutput("CommittedAgWaterUse", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
