% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcContGrazMax.R
\name{calcContGrazMax}
\alias{calcContGrazMax}
\title{calcContGrazMax}
\usage{
calcContGrazMax(report = "harvest")
}
\arguments{
\item{report}{Either 'harvest' or 'lsu/ha' controlling what values are output by the function.}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates continuous grazing maximum output
}
\examples{
\dontrun{calcOutput("ContGrazMax", report = "harvest")}

}
\author{
Marcos Alves
}
