% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcContGrazMax_new.R
\name{calcContGrazMax_new}
\alias{calcContGrazMax_new}
\title{calcContGrazMax_new}
\usage{
calcContGrazMax_new(
  lsu_levels = c(seq(0, 2.2, 0.2), 2.5),
  lpjml = "lpjml5.2_pasture",
  climatetype = "IPSL_CM6A_LR",
  scenario = "ssp126_co2_limN",
  report = "harvest"
)
}
\arguments{
\item{lsu_levels}{Livestock unit levels in the source folder}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{scenario}{specify ssp scenario}

\item{report}{Either 'harvest' or 'lsu/ha' controlling what values are output by the function.}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates continuous grazing maximum output
}
\examples{
\dontrun{calcOutput("ContGrazMax_new", lsu_levels = 0, lpjml, climatetype, report)}

}
\author{
Marcos Alves
}
