% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnvironment.R
\name{calcEnvironment}
\alias{calcEnvironment}
\title{calcEnvironment}
\usage{
calcEnvironment(climatetype = "HadGEM2_ES:rcp8p5:co2", sar = 20)
}
\arguments{
\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{sar}{Average range for smoothing annual variations}
}
\value{
magpie object in cellular resolution
}
\description{
Calculate climate, CO2 and soil environmental conditions on cellular level
}
\examples{
\dontrun{
calcOutput("Environment", climatetype = "HadGEM2_ES:rcp8p5:co2", sar = 20)
}

}
\author{
Marcos Alves
}
