% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnvmtlFlow.R
\name{calcEnvmtlFlow}
\alias{calcEnvmtlFlow}
\title{calcEnvmtlFlow}
\usage{
calcEnvmtlFlow(
  lpjml = "lpjml5.9.5-m1",
  climatetype = "GSWP3-W5E5:historical",
  seasonality = "grper"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{seasonality}{grper (default): EFR in growing period per year; total:
EFR throughout the year; monthly: monthly EFRs}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates environmental flow requirements (EFR) for MAgPIE
             retrieved from LPJmL monthly discharge and water availability
}
\examples{
\dontrun{
calcOutput("EnvmtlFlow", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
