% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEvapotranspiration.R
\name{calcEvapotranspiration}
\alias{calcEvapotranspiration}
\title{calcEvapotranspiration}
\usage{
calcEvapotranspiration(subtype = "H08:mri-esm2-0")
}
\arguments{
\item{subtype}{type of climate data to collect}
}
\value{
magpie object in cellular resolution
}
\description{
Disaggregate CO2 global atmospheric concentration to cellular level
}
\examples{
\dontrun{
calcOutput("GCMClimate_new", subtype = "ISIMIP3b:IPSL-CM6A-LR:ssp126:1850-2100:tas")
}

}
\author{
Marcos Alves, Kristine Karstens
}
