% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGCMClimate.R
\name{calcGCMClimate}
\alias{calcGCMClimate}
\title{calcGCMClimate}
\usage{
calcGCMClimate(
  subtype = "ISIMIP3bv2:IPSL-CM6A-LR:ssp126:1850-2100:tas:annual_mean",
  smooth = 0,
  cells = "lpjcell"
)
}
\arguments{
\item{subtype}{type of climate data to collect,
consisting of data source, GDM, RCP, time period, variable and time resolution
separated by ":"}

\item{smooth}{set averaging value for smoothing trajectories}

\item{cells}{number of halfdegree grid cells to be returned.
Options: "magpiecell" (59199), "lpjcell" (67420)}
}
\value{
magpie object in cellular resolution
}
\description{
Disaggregate CO2 global atmospheric concentration to cellular level
             NOTE: This function will be depreciate soon, please use mrland::calcLPJmLClimate
}
\examples{
\dontrun{
calcOutput("GCMClimate", subtype = "ISIMIP3b:IPSL-CM6A-LR:ssp126:1850-2100:tas:annual_mean")
}

}
\author{
Marcos Alves, Kristine Karstens, Felicitas Beier
}
