% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGCMClimate_new.R
\name{calcGCMClimate_new}
\alias{calcGCMClimate_new}
\title{calcGCMClimate_new}
\usage{
calcGCMClimate_new(
  subtype = "ISIMIP3b:IPSL-CM6A-LR:ssp126:1850-2100:tas",
  smooth = 0
)
}
\arguments{
\item{subtype}{type of climate data to collect}

\item{smooth}{set averaging value for smoothing trajectories}
}
\value{
magpie object in cellular resolution
}
\description{
Disaggregate CO2 global atmospheric concentration to cellular level
}
\examples{
\dontrun{
calcOutput("GCMClimate_new", subtype = "ISIMIP3b:IPSL-CM6A-LR:ssp126:1850-2100:tas")
}

}
\author{
Marcos Alves, Kristine Karstens
}
