% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrasslandBiomass.R
\name{calcGrasslandBiomass}
\alias{calcGrasslandBiomass}
\title{calcGrasslandBiomass}
\usage{
calcGrasslandBiomass(cells = "lpjcell")
}
\arguments{
\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
Regional biomass demand
}
\description{
Calculates pasture biomass demand for the historical period split
between rangelands andmanaged pastures.
}
\examples{
\dontrun{
calcOutput("GrasslandBiomass")
}
}
\seealso{
\code{\link[madrat]{calcOutput}}, \code{\link[mrcommons]{calcFAOmassbalance}},
\code{\link[madrat]{readSource}}
}
\author{
Marcos Alves, Bin Lin
}
