% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrasslandsYields.R, R/calcPastYields_new.R
\name{calcGrasslandsYields}
\alias{calcGrasslandsYields}
\title{calcGrasslandsYields}
\usage{
calcGrasslandsYields(
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  past_mngmt = "me2",
  lpjml = "lpjml5p2_pasture",
  climatetype = "IPSL_CM6A_LR",
  scenario = "ssp126_co2_limN"
)

calcGrasslandsYields(
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  past_mngmt = "me2",
  lpjml = "lpjml5p2_pasture",
  climatetype = "IPSL_CM6A_LR",
  scenario = "ssp126_co2_limN"
)
}
\arguments{
\item{lsu_levels}{Livestock unit levels in the source folder}

\item{past_mngmt}{pasture areas management option}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{scenario}{specify ssp scenario}
}
\value{
magpie object in cellular resolution

magpie object in cellular resolution
}
\description{
Calculates rangelands maximum output and managed pastures yields

Calculates rangelands maximum output and managed pastures yields
}
\examples{
\dontrun{
calcOutput("GrasslandsYields", lsu_levels, past_mngmt = "2me", lpjml, climatetype)
}

\dontrun{
calcOutput("GrasslandsYields", lsu_levels, past_mngmt = "2me", lpjml, climatetype)
}

}
\author{
Marcos Alves

Marcos Alves
}
