% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrassLndYldHist.R
\name{calcGrassLndYldHist}
\alias{calcGrassLndYldHist}
\title{Calculate historical grasslands yields split between pastures and rangelands weighted by livestock densities}
\usage{
calcGrassLndYldHist(cellular = TRUE, max_yields = 20, max_iter = 30)
}
\arguments{
\item{cellular}{Boolean to indicate if the results should be aggregated on regional or celullar level}

\item{max_yields}{Maximum yields in tDM/ha allowed in a cell.}

\item{max_iter}{Maximum number of iterations of the disaggregation algorithm}
}
\value{
Pasture yields and corresponding weights as a list of
two MAgPIE objects
}
\description{
`max_yields` and `max_iter` are only affecting the calculations if `cellular` is TRUE.
}
\details{
Provides Pasture yields defined as ratio of removed biomass to grassland area
}
\examples{
\dontrun{
calcOutput("GrassLndYldHist", max_yields = 20, max_iter = 30)
}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{calcFAOmassbalance}},
\code{\link{readSource}}
}
\author{
Marcos Alves
}
