% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrassPastureShare.R
\name{calcGrassPastureShare}
\alias{calcGrassPastureShare}
\title{calcGrassPastureShare}
\usage{
calcGrassPastureShare()
}
\value{
List of magpie object with results on cluster level
}
\description{
Calculate glassland shareas os pasture managed lands.
}
\examples{

\dontrun{
calcOutput("GrassPastureShare")
}
}
\author{
Marcos Alves
}
