% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrassyEcoregions.R
\name{calcGrassyEcoregions}
\alias{calcGrassyEcoregions}
\title{calcGrassyEcoregions}
\usage{
calcGrassyEcoregions()
}
\value{
magpie object in cellular resolution (67420 lpjcells)
}
\description{
Per-cell areal cover fraction of the RESOLVE Ecoregions 2017 grassy biomes
(Dinerstein et al. 2017): tropical and subtropical, temperate, flooded, and montane
grasslands, savannas and shrublands (biomes 7-10). Used to correct the LPJmL-derived
potential forest area, which overestimates forest cover in these open ecosystems
(Veldman et al. 2015).
}
\examples{
\dontrun{
calcOutput("GrassyEcoregions", aggregate = FALSE)
}

}
\author{
Florian Humpenoeder
}
