% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGridPop_new.R
\name{calcGridPop_new}
\alias{calcGridPop_new}
\title{calcGridPop_new}
\usage{
calcGridPop_new(
  subtype = "all",
  cellular = TRUE,
  FiveYear = TRUE,
  harmonize_until = 2015
)
}
\arguments{
\item{subtype}{past (1965-2005), future (2005-2010) or all (divergence starts at year in harmonize_until)}

\item{cellular}{only cellular}

\item{FiveYear}{TRUE for 5 year time steps, otherwise yearly from source}

\item{harmonize_until}{2015 default divergence of SSPs}
}
\value{
Population in millions.
}
\description{
Past and future (SSP1-5) population based on HYDE3.2 and Jones & O'Neill (2016)
Data is scaled to match WDI data from calcPopulation
NOTE that some scaling factors for the future (for small countries Gambia and Djibouti) are off,
data read in is 50% of WDI data, most likely due to large resolution
}
\author{
David Chen
}
