% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigatableArea.R
\name{calcIrrigatableArea}
\alias{calcIrrigatableArea}
\title{calcIrrigatableArea}
\usage{
calcIrrigatableArea(cells = "lpjcell")
}
\arguments{
\item{cells}{Switch between "lpjcell" (67420) and "magpiecell" (59199)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates the area that can be irrigated given available area and available water for all magpie crops
}
\examples{
\dontrun{ calcOutput("IrrigatableArea",aggregate=FALSE) }

}
\author{
Felicitas Beier
}
