% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigation2.R
\name{calcIrrigation2}
\alias{calcIrrigation2}
\title{calcIrrigation2}
\usage{
calcIrrigation2(
  selectyears = "all",
  cells = "lpjcell",
  version = "LPJmL5",
  climatetype = "CRU_4",
  time = "raw",
  averaging_range = NULL,
  dof = NULL,
  harmonize_baseline = FALSE,
  ref_year = NULL
)
}
\arguments{
\item{selectyears}{years to be returned}

\item{cells}{Switch between "lpjcell" (67420) and "magpiecell" (59199)}

\item{version}{Switch between LPJmL4 and LPJmL5}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{Time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{harmonize_baseline}{FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{Reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates irrigation water withdrawal based on LPJmL plant blue water consumption and different irrigation efficiencies
}
\examples{
\dontrun{ calcOutput("Irrigation", aggregate = FALSE) }

}
\author{
Felicitas Beier, Jens Heinke
}
