% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISIMIP3bYields.R
\name{calcISIMIP3bYields}
\alias{calcISIMIP3bYields}
\title{calcISIMIP3bYields}
\usage{
calcISIMIP3bYields(
  subtype = "yields:EPIC-IIASA:ukesm1-0-ll:ssp585:default:3b",
  cells = "magpiecells",
  time = "spline",
  dof = 4
)
}
\arguments{
\item{subtype}{subtype of yield based on readISIMIPoutputs, for crop yields}

\item{cells}{magpie or lpjcell
@return magpie object in cellular resolution}

\item{time}{spline or average, if spline specify dof, if average specify averaging range}

\item{dof}{degrees of freedom for spline}
}
\description{
reads and cleans up ISIMIP3b crop yield data
}
\examples{
\dontrun{ calcOutput("ISIMIP3bYields", aggregate = FALSE) }

}
\author{
David Meng-Chuen Chen
}
