% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLabourProdImpactEmu.R
\name{calcLabourProdImpactEmu}
\alias{calcLabourProdImpactEmu}
\title{calcLabourProdImpactEmu}
\usage{
calcLabourProdImpactEmu(timestep = "5year", cellular = TRUE)
}
\arguments{
\item{timestep}{5-year or yearly}

\item{cellular}{cellular is true}
}
\value{
List of magpie object of gridded (0.5) productivity as share of 1 (full productivity)
}
\description{
Spatial and temporal aggr. of labour productivity impacts from climate change emulated by LAMACLIMA

based on method of Orlov et al. 2019. Economics of Disasters and Climate Change, 3(3), 191-211.
}
\author{
Michael Windisch
}
