% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLivestockDistribution.R
\name{calcLivestockDistribution}
\alias{calcLivestockDistribution}
\title{calcLivestockDistribution}
\usage{
calcLivestockDistribution(cells = "lpjcell")
}
\arguments{
\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
MAgPIE objects with livestock numbers on a cellular level.
}
\description{
Disaggregate Livestock estimates based on the GLW3 dataset.
}
\examples{
\dontrun{
calcOutput("LivestockDistribution")
}

}
\author{
Marcos Alves
}
