% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLsuDensityHist.R
\name{calcLsuDensityHist}
\alias{calcLsuDensityHist}
\title{calcLsuDensityHist}
\usage{
calcLsuDensityHist(disagg_type = "grassland", cells = "lpjcell")
}
\arguments{
\item{disagg_type}{select the disaggregaton weights for biomass production (can be either grassland or livestock)}

\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
List of magpie object with results on cluster level
}
\description{
Calculate livestock historical livestock densities
}
\examples{
\dontrun{
calcOutput("LsuDensityHist")
}
}
\author{
Marcos Alves
}
