% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMAPSPAM.R
\name{calcMAPSPAM}
\alias{calcMAPSPAM}
\title{calcMAPSPAM}
\usage{
calcMAPSPAM(subtype = "physical")
}
\arguments{
\item{subtype}{it can be either "physical" or "harvested" area}
}
\value{
magpie object in cellular resolution
}
\description{
MAPSPAM data
}
\examples{
\dontrun{
calcOutput("MAPSPAM", subtype = "physical", aggregate = FALSE)
}

}
\author{
Edna J. Molina Bacca
}
