% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMowing.R
\name{calcMowing}
\alias{calcMowing}
\title{calcMowing}
\usage{
calcMowing()
}
\value{
magpie object in cellular resolution
}
\description{
Calculates mowing pasture yields
}
\examples{
\dontrun{calcOutput("Mowing")}

}
\author{
Marcos Alves
}
