% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMowing_new.R
\name{calcMowing_new}
\alias{calcMowing_new}
\title{calcMowing_new}
\usage{
calcMowing_new(
  mowing_events = "me2",
  lpjml = "lpjml5.2_pasture",
  climatetype = "IPSL_CM6A_LR",
  scenario = "ssp126_co2_limN"
)
}
\arguments{
\item{mowing_events}{number of mowing events per year expressed as `2me`}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{scenario}{specify ssp scenario}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates mowing pasture yields
}
\examples{
\dontrun{
calcOutput("Mowing_new", mowing_events = "2me", lpjml = "LPJml_mowing", climatetype)
}

}
\author{
Marcos Alves
}
