% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonAgWaterDemand.R
\name{calcNonAgWaterDemand}
\alias{calcNonAgWaterDemand}
\title{calcNonAgWaterDemand}
\usage{
calcNonAgWaterDemand(
  selectyears = "all",
  source = "WATCH_ISIMIP_WATERGAP",
  time = "raw",
  averaging_range = NULL,
  dof = NULL,
  waterusetype = "withdrawal",
  seasonality = "total"
)
}
\arguments{
\item{selectyears}{years to be returned}

\item{source}{data source to be used}

\item{time}{Time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{waterusetype}{withdrawal (default) or consumption}

\item{seasonality}{total (default) or grper (growing period)}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts non-agricultural water demand
}
\examples{
\dontrun{ calcOutput("NonAgWaterDemand", aggregate = FALSE) }

}
\author{
Felicitas Beier
}
