% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonLocalProduction.R
\name{calcNonLocalProduction}
\alias{calcNonLocalProduction}
\title{calcNonLocalProduction}
\usage{
calcNonLocalProduction(cells = "lpjcell")
}
\arguments{
\item{cells}{magpiecell or lpjcell (default)}
}
\description{
Calculates grid-level amount of food that would need to be transported,
assuming that food produced in the grid cell is first consumed by local population
i.e. amount of food greater than local rural demand, split into that which feeds the local
urban population, and that which exceeds total local demand and is available to export
}
\examples{
\dontrun{
calcOutput("NonLocalTransport")
}
}
\author{
David M Chen
}
