% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonLocalTransport.R
\name{calcNonLocalTransport}
\alias{calcNonLocalTransport}
\title{calcNonLocalTransport}
\usage{
calcNonLocalTransport()
}
\description{
Calculates grid-level amount of food that would be transported,
assuming that food produced in the grid cell is first consumed by local population
i.e. amount of food greater than local rural demand
}
\examples{
\dontrun{
calcOutput("NonLocalTransport")
}
}
\author{
David M Chen
}
