% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPackagingMarketingCosts.R
\name{calcPackagingMarketingCosts}
\alias{calcPackagingMarketingCosts}
\title{calcPackagingMarketingCosts}
\usage{
calcPackagingMarketingCosts()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates per-ton marketing and packaging costs for food that leaves a cell
Currnetly assume expert guess 50 USD / ton of packaging/marketing costs 
(100 USD/t in model, of which half is already in GTAP)
}
\author{
David M Chen
}
