% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastr.R
\name{calcPastr}
\alias{calcPastr}
\title{calcPastr}
\usage{
calcPastr(
  past_mngmt = "me2",
  lpjml = "LPJmL_pastr",
  climatetype = "HadGEM2_ES:rcp8p5:co2"
)
}
\arguments{
\item{past_mngmt}{pasture areas management option}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates managed pastures yields
}
\examples{
\dontrun{
calcOutput("Pastr", past_mngmt = "me2", lpjml = "LPJmL_pastr", climatetype)
}

}
\author{
Marcos Alves
}
