% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastr_new.R
\name{calcPastr_new}
\alias{calcPastr_new}
\title{calcPastr_new}
\usage{
calcPastr_new(
  past_mngmt = "me2",
  lpjml = "lpjml5p2_pasture",
  climatetype = "IPSL_CM6A_LR",
  scenario = "ssp126_co2_limN",
  cells = "lpjcell"
)
}
\arguments{
\item{past_mngmt}{pasture areas management option}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{scenario}{specify ssp scenario}

\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates managed pasture yields
}
\examples{
\dontrun{
calcOutput("Pastr_new", past_mngmt = "me2", lpjml = "LPJml_pastr", climatetype)
}

}
\author{
Marcos Alves
}
