% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastrMngtLevels.r
\name{calcPastrMngtLevels}
\alias{calcPastrMngtLevels}
\title{calcPastrMngtLevels}
\usage{
calcPastrMngtLevels(
  climatetype = "MRI-ESM2-0:ssp370",
  options = c("brazil_1", "brazil_2", "brazil_4"),
  cost_level = c(1, 2, 3),
  cells = "lpjcell"
)
}
\arguments{
\item{climatetype}{SSP+RCP combination}

\item{options}{Management options simulated by LPJml}

\item{cost_level}{level cost for different past management options}

\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
magpie object in 0.5 degree resolution
}
\description{
Calculates managed pasture potential yields for different combinations of SSP+RCP and grassland management options
}
\examples{
\dontrun{
calcOutput("PastrMngtLevels", ssps, options)
}

}
\author{
Marcos Alves
}
