% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastureSuitOld.R
\name{calcPastureSuitOld}
\alias{calcPastureSuitOld}
\title{calcPastureSuit}
\usage{
calcPastureSuitOld(
  climatetype = "MRI-ESM2-0:ssp126",
  lpjml = "LPJmL4_for_MAgPIE_44ac93de"
)
}
\arguments{
\item{climatetype}{Switch between different climate scenarios}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}
}
\value{
List of magpie object with results on cluster level
}
\description{
Calculate glassland suitable for pasture management based
on population and aridity criteria.
}
\examples{
\dontrun{
calcOutput("PastureSuitOld")
}
}
\author{
Marcos Alves
}
