% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastYields_new.R
\name{calcPastYields_new}
\alias{calcPastYields_new}
\title{calcPastYields_new}
\usage{
calcPastYields_new(
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  mowing_events = "me2",
  lpjml = "lpjml5.2_pasture",
  climatetype = "IPSL_CM6A_LR",
  scenario = "ssp126_co2_limN"
)
}
\arguments{
\item{lsu_levels}{Livestock unit levels in the source folder}

\item{mowing_events}{number of mowing events per year expressed as `2me`}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{scenario}{specify ssp scenario}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates continuous grazing maximum output and mowing yields
}
\examples{
\dontrun{calcOutput("PastYields_new", lsu_levels, mowing_events = "2me", lpjml, climatetype)}

}
\author{
Marcos Alves
}
