% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPeatland2.R
\name{calcPeatland2}
\alias{calcPeatland2}
\title{calcPeatland2}
\usage{
calcPeatland2(cells = "magpiecell")
}
\arguments{
\item{cells}{number of cells to be returned: magpiecell (59199), lpjcell (67420)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates degraded and intact peatland area at cell level.
The function takes degraded and intact peatland area from the Global Peatland Database 2022 (GPD2022) at the
national level and downscales the peatland area to grid cell level using gridded peatland area
from the Global Peatland Map 2.0 (GPM2)
The data has been provided by Alexandra Barthelmes.
}
\examples{
\dontrun{
calcOutput("Peatland2", aggregate = FALSE)
}

}
\author{
Florian Humpenoeder
}
