% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRangelandsMax.R
\name{calcRangelandsMax}
\alias{calcRangelandsMax}
\title{calcRangelandsMax}
\usage{
calcRangelandsMax(
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  lpjml = "LPJmL_range",
  climatetype = "HadGEM2_ES:rcp8p5:co2",
  report = "harvest"
)
}
\arguments{
\item{lsu_levels}{Livestock unit levels in the source folder}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{report}{Either 'harvest' or 'lsu/ha' controlling what values are output by the function.}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates rangelands maximum output
}
\examples{
\dontrun{calcOutput("ContGrazMax", lsu_levels = 0, lpjml, climatetype, report)}

}
\author{
Marcos Alves
}
