% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRangelandsMaxNew.R
\name{calcRangelandsMaxNew}
\alias{calcRangelandsMaxNew}
\title{calcRangelandsMaxNew}
\usage{
calcRangelandsMaxNew(
  lsuLevels = c(seq(0, 2.2, 0.2), 2.5),
  lpjml = "lpjml5p2_pasture",
  climatetype = "MRI-ESM2-0:ssp370",
  scenario = "/co2/Nreturn0p5/limN",
  report = "harvest",
  cells = "lpjcell"
)
}
\arguments{
\item{lsuLevels}{Livestock unit levels in the source folder}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{scenario}{specify ssp scenario}

\item{report}{Either 'harvest' or 'lsu/ha' controlling what values are output by the function.}

\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates rangelands maximum output
}
\examples{
\dontrun{
calcOutput("ContGrazMax_new", lsuLevels = 0, lpjml, climatetype, report)
}

}
\author{
Marcos Alves
}
