% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRangeSoilCarbonHist.R
\name{calcRangeSoilCarbonHist}
\alias{calcRangeSoilCarbonHist}
\title{calcRangeSoilCarbonHist}
\usage{
calcRangeSoilCarbonHist(
  subtype = "ISIMIP3b:IPSL-CM6A-LR:ssp126:1965-2100",
  lpjml,
  model = "9eaf9b"
)
}
\arguments{
\item{subtype}{subtypes}

\item{lpjml}{lpjml version}

\item{model}{trained model ID}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates soil carbon for rangelands
}
\examples{
\dontrun{
calcOutput("GrassSoilCarbonHist ", subtype, model)
}
}
\author{
Marcos Alves
}
