% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcScaleEnvironmentData_new.R
\name{calcScaleEnvironmentData_new}
\alias{calcScaleEnvironmentData_new}
\title{calcScaleEnvironmentData_new}
\usage{
calcScaleEnvironmentData_new(
  subtype = "ISIMIP3b:IPSL-CM6A-LR:ssp126:1965-2100",
  aggr = F,
  sar = 20,
  sel_feat = c("tas", "pr", "lwnet", "rsds", "CO2", "Ks", "Sf", "w_pwp", "w_fc",
    "w_sat", "hsg")
)
}
\arguments{
\item{subtype}{Switch between different climate scenarios}

\item{aggr}{aggregation level}

\item{sar}{Average range for smoothing annual variations}

\item{sel_feat}{features names to be included in the output file}
}
\value{
magpie object in cellular resolution
}
\description{
Scale climate, CO2 and soil environmental conditions on cellular level
}
\examples{
\dontrun{
calcOutput("ScaleEnvironmentData_new", climatetype = "HadGEM2_ES:rcp8p5:co2", sar = 20, sel_feat)
}

}
\author{
Marcos Alves
}
