% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSchulteUebbing.R
\name{calcSchulteUebbing}
\alias{calcSchulteUebbing}
\title{calcSchulteUebbing}
\usage{
calcSchulteUebbing()
}
\value{
a magpie object containing critical nitrogen surplus
}
\description{
use inverse distance weighted interpolation on critical nitrogen surplus data set of
Schulte-Uebbing et al. 2022
}
\examples{

\dontrun{
  calcOutput("SchulteUebbing")
}
}
\author{
Michael Crawford
}
