% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSCScalingFactors.R
\name{calcSCScalingFactors}
\alias{calcSCScalingFactors}
\title{calcSCScalingFactors}
\usage{
calcSCScalingFactors(
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  lpjml = "LPJML5.2_pasture",
  climatetype = "IPSL_CM6A_LR",
  scenario = "ssp126_co2_limN",
  sar = 20
)
}
\arguments{
\item{lsu_levels}{Livestock unit levels in the source folder}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{scenario}{scenario specifications (eg. ssp126_co2_limN)}

\item{sar}{Average range for smoothing annual variations}
}
\value{
magpie object in cellular resolution
}
\description{
calculates the mean and sd of the scaled pasture soil carbon dataset
}
\examples{
\dontrun{
calcOutput("SCScalingFactors", lsu_levels = c(seq(0, 2, 0.2), 2.5), scenario)
}



}
\author{
Marcos Alves
}
