% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSoilcarbonLayer.R
\name{calcSoilcarbonLayer}
\alias{calcSoilcarbonLayer}
\title{calcSoilcarbonLayer}
\usage{
calcSoilcarbonLayer(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_44ac93de", crop =
    "ggcmi_phase3_nchecks_9ca735cb"),
  climatetype = "MRI-ESM2-0:ssp370",
  weight = "totalPNV"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different GCM climate scenarios}

\item{weight}{Switch between different weights:
'totalPNV' for LandArea as weight,
'PotForest' for PotentialForestArea}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts soil carbon densities from LPJmL for PNV to MAgPIE
}
\examples{
\dontrun{
calcOutput("calcSoilcarbonLayer", aggregate = FALSE)
}

}
\author{
Kristine Karstens
}
