% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSoilTraining.R
\name{calcSoilTraining}
\alias{calcSoilTraining}
\title{calcSoilCarbon}
\usage{
calcSoilTraining(
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  lpjml = "LPJmL_cgrazing",
  climatetype = "HadGEM2_ES:rcp8p5:co2",
  sar = 20
)
}
\arguments{
\item{lsu_levels}{Livestock unit levels in the source folder}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{sar}{Smooth average range}
}
\value{
magpie object in cellular resolution
}
\description{
Calculate soil carbon stocks for different LSU and climate conditions
}
\examples{
\dontrun{
calcOutput("SoilTraining")
}

}
\author{
Marcos Alves
}
