% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopsoilCarbon.R
\name{calcTopsoilCarbon}
\alias{calcTopsoilCarbon}
\title{calcTopsoilCarbon}
\usage{
calcTopsoilCarbon(lpjml = "lpjml5.9.5-m1", climatetype = "MRI-ESM2-0:ssp370")
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different GCM climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts topsoil carbon densities from LPJ to MAgPIE
}
\examples{
\dontrun{
calcOutput("TopsoilCarbon", aggregate = FALSE)
}
}
\author{
Kristine Karstens
}
