% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTopsoilCarbon_new.R
\name{calcTopsoilCarbon_new}
\alias{calcTopsoilCarbon_new}
\title{calcTopsoilCarbon_new}
\usage{
calcTopsoilCarbon_new(
  lpjml = c(natveg = "LPJmL4", crop = "LPJmL5"),
  climatetype
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different GCM climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts topsoil carbon densities from LPJ to MAgPIE
}
\examples{
\dontrun{ calcOutput("TopsoilCarbon_new", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
