% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransportCosts.R
\name{calcTransportCosts}
\alias{calcTransportCosts}
\title{calcTransportCosts}
\usage{
calcTransportCosts(transport = "all", gtapVersion = "9")
}
\arguments{
\item{transport}{"all" or "nonlocal". "all" means all production incurs transport costs,
while "nonlocal" sees only production greater than local rural consumption with transport costs}

\item{gtapVersion}{"9" or "81"}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates country-level transport costs from GTAP total
transport costs, cellular production, and cellular travel time
}
\examples{
\dontrun{
calcOutput("TransportCosts_new")
}

}
\seealso{
[calcTransportTime()],
[calcGTAPTotalTransportCosts()]
}
\author{
David M Chen
}
