% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransportCosts_new.R
\name{calcTransportCosts_new}
\alias{calcTransportCosts_new}
\title{calcTransportCosts_new}
\usage{
calcTransportCosts_new()
}
\description{
calculates country-level transport costs from GTAP total
transport costs, cellular production, and cellular travel time

#' @return List of magpie objects with results on country level, weight on country level, unit and description.
}
\examples{
\dontrun{
calcOutput("TransportCosts_new")
}
}
\seealso{
[calcTransportTime()],
[calcGTAPTotalTransportCosts()]
}
\author{
David M Chen
}
