% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransportDistance.R
\name{calcTransportDistance}
\alias{calcTransportDistance}
\title{calcTransportDistance}
\usage{
calcTransportDistance()
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts travel time to major cities in minutes
This function now deprecated - use calcTransportTime instead
}
\examples{
\dontrun{
calcOutput("TransportDistance", aggregate = FALSE)
}

}
\author{
David Chen
}
