% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransportTime.R
\name{calcTransportTime}
\alias{calcTransportTime}
\title{calcTransportTime}
\usage{
calcTransportTime(object = "cities", minDist = 5)
}
\arguments{
\item{object}{cities or ports}

\item{minDist}{5 or 50 for cities, Large|Medium|Small|Very Small|Any for ports. For Cities, indicates
5 000 or 20 000 or 50 000 as settlement population size for travel time to, while .}
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts travel time to major cities in minutes
}
\examples{
\dontrun{ calcOutput("TransportTime", aggregate = FALSE) }


}
\author{
David Chen
}
