% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTRENDY.R
\name{calcTRENDY}
\alias{calcTRENDY}
\title{calcTRENDY}
\usage{
calcTRENDY(subtype)
}
\arguments{
\item{subtype}{name of the TRENDY dataset to read. Current options includes:
"CABLEPOP", "CARDAMOM", "CLASSIC", "CLM5.0", "DLEM",
"ELM", "IBIS", "ISBACTRIP", "JSBACH", "LPJ-GUESS",
"LPJml", "LPJwsl", "lpxqs", "OCN", "ORCHIDEE",
"SDGVM", "VISIT"}
}
\value{
a magclass object with model, carbon pool in third dimension
}
\description{
Resample and use inverse distance weighted interpolation to align
             natural vegetation carbon densities from TRENDY with MAgPIE's input requirements
}
\examples{
\dontrun{
readSource("TRENDY", subtype = "JSBACH")
}

}
\author{
Michael Crawford
}
