% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterUseNonAg.R
\name{calcWaterUseNonAg}
\alias{calcWaterUseNonAg}
\title{calcWaterUseNonAg}
\usage{
calcWaterUseNonAg(
  selectyears = "all",
  source = "WATCH_ISIMIP_WATERGAP",
  seasonality = "grper",
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_84a69edd", crop =
    "ggcmi_phase3_nchecks_72c185fa"),
  climatetype = "GSWP3-W5E5:historical"
)
}
\arguments{
\item{selectyears}{years to be returned}

\item{source}{data source to be used (e.g. WATERGAP2020)}

\item{seasonality}{grper (default): non-agricultural water demand in growing period per year; total: non-agricultural water demand throughout the year}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{switch between different climate scenarios (default: "CRU_4") for calcGrowingPeriod}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts non-agricultural water demand
}
\examples{
\dontrun{ calcOutput("WaterUseNonAg", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
