% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldImprovementPotential.R
\name{calcYieldImprovementPotential}
\alias{calcYieldImprovementPotential}
\title{calcYieldImprovementPotential}
\usage{
calcYieldImprovementPotential(
  version = "LPJmL5",
  climatetype = "CRU_4",
  time = "spline",
  averaging_range = NULL,
  dof = 4,
  harmonize_baseline = FALSE,
  selectyears = seq(1995, 2095, by = 5),
  cells = "magpiecell"
)
}
\arguments{
\item{version}{switch between LPJmL4 and LPJmL5}

\item{climatetype}{switch between different climate scenarios (default: "CRU_4")}

\item{time}{time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{harmonize_baseline}{FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{selectyears}{defaults to all years available}

\item{cells}{switch between "lpjcell" (67420) and "magpiecell" (59199)}

\item{ref_year}{reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}
}
\value{
magpie object in cellular resolution
}
\description{
This calculates the yield improvement potential of irrigation for different crops
}
\examples{
\dontrun{ calcOutput("YieldImprovementPotential", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
