% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYields_new.R
\name{calcYields_new}
\alias{calcYields_new}
\title{calcYields_new}
\usage{
calcYields_new(
  lpjml = c(natveg = "LPJmL4", crop = "LPJmL5"),
  climatetype = "GSWP3-W5E5:historical",
  cells = "magpiecell"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{cells}{if cellular is TRUE: "magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts yields from LPJmL to MAgPIE
}
\examples{
\dontrun{ calcOutput("Yields", aggregate = FALSE) }

}
\author{
Kristine Karstens, Felicitas Beier
}
