% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldsAggregationWeights.R
\name{calcYieldsAggregationWeights}
\alias{calcYieldsAggregationWeights}
\title{calcYieldsAggregationWeights}
\usage{
calcYieldsAggregationWeights(
  weighting = "totalCrop",
  lpjml = NULL,
  climatetype = NULL,
  multicropping = NULL,
  marginal_land = "q33_marginal:rainfed_and_irrigated"
)
}
\arguments{
\item{weighting}{area-based weighting options for yield aggregation
("totalCrop" (default),
"totalLUspecific", "cropSpecific", "crop+irrigSpecific",
"avlCropland", "avlCropland+potentiallyIrrigatedAreas",
"avlCropland+avlPasture")}

\item{lpjml}{lpjml version, only required if potentially irrigated areas
are used as aggregation weight}

\item{climatetype}{different climate scenarios, only required if potentially
irrigated areas are used as aggregation weight}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected
(e.g. TRUE:actual:total; TRUE:none; FALSE)}

\item{marginal_land}{Defines which share of marginal land should be included
if avlCropland is chosen as weighting option.}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates the crop area weightings to use for yield aggregation.
}
\examples{
\dontrun{
calcOutput("YieldsAggregationWeights", aggregate = FALSE)
}

}
\author{
Kristine Karstens, Felicitas Beier
}
