% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGPD2022.R
\name{convertGPD2022}
\alias{convertGPD2022}
\title{convertGPD2022}
\usage{
convertGPD2022(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on iso level, weight, unit and description.
}
\description{
convert GPD2022
}
\examples{
\dontrun{
readSource("GPD2022", convert = TRUE)
}

}
\author{
Florian Humpenoeder
}
