% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBphMask.R
\name{correctBphMask}
\alias{correctBphMask}
\title{correctBphMask}
\usage{
correctBphMask(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie objects on cellular level
}
\description{
correct cellnames
}
\examples{

\dontrun{
  readSource("readBphMask", convert="onlycorrect")
}
}
\author{
Kristine Karstens
}
