% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctDinerstein2017.R
\name{correctDinerstein2017}
\alias{correctDinerstein2017}
\title{correctDinerstein2017}
\usage{
correctDinerstein2017(x)
}
\arguments{
\item{x}{magpie object provided by \code{\link{readDinerstein2017}}}
}
\value{
magpie object on cellular level
}
\description{
Sets missing values (cells outside any ecoregion of the selected biome
group) to zero.
}
\examples{
\dontrun{
readSource("Dinerstein2017", subtype = "grassland", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readDinerstein2017}}
}
\author{
Florian Humpenoeder
}
