% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGAMI.R
\name{correctGAMI}
\alias{correctGAMI}
\title{correctGAMI}
\usage{
correctGAMI(x)
}
\arguments{
\item{x}{magpie object provided by \code{readGAMI}}
}
\value{
magpie object on the 67420 lpj-cell grid, cleaned
}
\description{
Correct the GAMI v2.1 forest-age dataset: replace missing / negative
class-fraction values by 0 and attach ISO country codes to the coordinate cells.
}
\examples{
\dontrun{
readSource("GAMI", convert = "onlycorrect")
}
}
\seealso{
\code{\link{readGAMI}}
}
\author{
Florian Humpenoeder
}
