% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGCMClimate.R
\name{correctGCMClimate}
\alias{correctGCMClimate}
\title{correctGCMClimate}
\usage{
correctGCMClimate(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
Magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct GCMs climate variables
             NOTE: This function will be depreciate soon, please use mrland::correctLPJmLClimate
}
\examples{

\dontrun{
readSource("GCMClimate", subtype, convert="onlycorrect")
}

}
\seealso{
\code{\link{readGCMClimate}}
}
\author{
Marcos Alves, Felicitas Beier
}
