% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGCMClimate_new.R
\name{correctGCMClimate_new}
\alias{correctGCMClimate_new}
\title{correctGCMClimate_new}
\usage{
correctGCMClimate_new(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
Magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct GCMs climate variables
}
\examples{

\dontrun{
readSource("GCMClimate_new", subtype, convert="onlycorrect")
}

}
\seealso{
\code{\link{readGCMClimate_new}}
}
\author{
Marcos Alves
}
