% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGPM2.R
\name{correctGPM2}
\alias{correctGPM2}
\title{correctGPM2}
\usage{
correctGPM2(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
correct peatland area
}
\examples{

\dontrun{
  readSource("GPM2", convert="onlycorrect")
}
}
\author{
Florian Humpenoeder
}
