% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctHalfEarth.R, R/correctZabel2014.R
\name{correctHalfEarth}
\alias{correctHalfEarth}
\title{correctHalfEarth}
\usage{
correctHalfEarth(x)

correctHalfEarth(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level

magpie object on cellular level
}
\description{
correct HalfEarth data

correct Zabel crop suitability data
}
\examples{

\dontrun{
  readSource("HalfEarth", convert="onlycorrect")
}


\dontrun{
  readSource("Zabel2014", convert="onlycorrect")
}

}
\seealso{
\code{\link{readHalfEarth}}

\code{\link{readZabel2014}}
}
\author{
Felicitas Beier

Patrick v. Jeetze
}
