% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLabourProdImpactEmu.R
\name{correctLabourProdImpactEmu}
\alias{correctLabourProdImpactEmu}
\title{correctLabourProdImpactEmu}
\usage{
correctLabourProdImpactEmu(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
correct labour productivity impacts from climate change emulated by the LAMACLIMA project

based on method of Orlov et al. 2019. Economics of Disasters and Climate Change, 3(3), 191-211.
}
\examples{

\dontrun{
  readSource("LabourProdImpactEmu", convert="onlycorrect")
}

}
\seealso{
\code{\link{readLabourProdImpactEmu}}
}
\author{
Michael Windisch
}
