% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctMehta2024.R
\name{correctMehta2024}
\alias{correctMehta2024}
\title{correctMehta2024}
\usage{
correctMehta2024(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object in cellular resolution
}
\description{
correct Global Area Equipped for Irrigation Dataset 1900-2015
             from Mehta et al., 2024
}
\examples{

\dontrun{
  readSource("Mehta2024", convert="onlycorrect")
}
}
\author{
Felicitas Beier
}
