% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctPastYields.R
\name{correctPastYields}
\alias{correctPastYields}
\title{correctPastYields}
\usage{
correctPastYields(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
Magpie objects with results on cellular level.
}
\description{
Correct pasture yields
}
\examples{

\dontrun{
  readSource("PastYields", subtype = "ContinousGrazing:elevatedCO2:rcp8p5.multiple")
}

}
\seealso{
\code{\link{readPastYields}}
}
\author{
Marcos Alves
}
