% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctRamankutty.R
\name{correctRamankutty}
\alias{correctRamankutty}
\title{correctRamankutty}
\usage{
correctRamankutty(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object
}
\description{
Read Available Land Si
}
\examples{

\dontrun{
  readSource("Ramankutty", convert="onlycorrect")
}

}
\seealso{
\code{\link{readRamankutty}}
}
\author{
Felicitas Beier
}
