% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCO2Atmosphere_new.R
\name{downloadCO2Atmosphere_new}
\alias{downloadCO2Atmosphere_new}
\title{downloadCO2Atmosphere_new}
\usage{
downloadCO2Atmosphere_new(subtype = "ISIMIP3b:ssp126")
}
\arguments{
\item{subtype}{Switch between different inputs (eg. "ISIMIP3b:IPSL-CM6A-LR:historical:1850-2014:tas")
It consists of GCM version, climate model, scenario and variable.}
}
\value{
metadata entry
}
\description{
Download CO2 atm. inputs used for Lpjml runs
}
\author{
Marcos Alves
}
