% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDinerstein2017.R
\name{downloadDinerstein2017}
\alias{downloadDinerstein2017}
\title{downloadDinerstein2017}
\usage{
downloadDinerstein2017()
}
\description{
Downloads the RESOLVE Terrestrial Ecoregions 2017 (Dinerstein et al. 2017),
a global map of 846 ecoregions grouped into 14 biomes.
}
\seealso{
\code{\link{readDinerstein2017}}
}
\author{
Florian Humpenoeder
}
