% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadEvapotranspiration.R
\name{downloadEvapotranspiration}
\alias{downloadEvapotranspiration}
\title{downloadEvapotranspiration}
\usage{
downloadEvapotranspiration(subtype = "H08:mri-esm2-0:historical")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\description{
Download water models evapotranspiration data
}
\examples{

\dontrun{readSource("Evapotranspiration",  convert="onlycorrect")}
}
\author{
Marcos Alves
}
