% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGAMI.R
\name{downloadGAMI}
\alias{downloadGAMI}
\title{downloadGAMI}
\usage{
downloadGAMI()
}
\value{
metadata list describing the downloaded GAMI source
}
\description{
Downloads the Global Age Mapping Integration (GAMI) v2.1 forest-age dataset
(Besnard et al. 2024, GFZ Data Services, doi:10.5880/GFZ.1.4.2023.006), the 0.5-degree
`class_fraction` product read by \code{\link{readGAMI}}. GAMI provides the within-forest
age distribution (12 age classes, 20-member ensemble) for 2010 and 2020.
}
\examples{
\dontrun{
downloadSource("GAMI")
}
}
\seealso{
[downloadSource()], \code{\link{readGAMI}}
}
\author{
Florian Humpenoeder
}
