% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGCMclimate.R
\name{downloadGCMClimate}
\alias{downloadGCMClimate}
\title{downloadGCMClimate}
\usage{
downloadGCMClimate(subtype = "ISIMIP3b:IPSL-CM6A-LR:ssp126:2015-2100:tas")
}
\arguments{
\item{subtype}{Switch between different inputs (e.g. "ISIMIP3b:IPSL-CM6A-LR:historical:1850-2014:tas")
Argument consists of GCM version, climate model, scenario and variable,
separated by ":"}
}
\value{
metadata entry
}
\description{
Download GCM climate input used for Lpjml runs
             NOTE: This function will be depreciate soon, please use mrland::downloadLPJmLClimate
}
\examples{
\dontrun{
readSource("GCMClimate", convert = "onlycorrect")
}

}
\author{
Marcos Alves
}
