% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGCMclimate_new.R
\name{downloadGCMClimate_new}
\alias{downloadGCMClimate_new}
\title{downloadGCMClimate_new}
\usage{
downloadGCMClimate_new(subtype = "ISIMIP3b:IPSL-CM6A-LR:ssp126:2015-2100:tas")
}
\arguments{
\item{subtype}{Switch between different inputs (eg. "ISIMIP3b:IPSL-CM6A-LR:historical:1850-2014:tas")
It consists of GCM version, climate model, scenario and variable.}
}
\value{
metadata entry
}
\description{
Download GCM climate input used for Lpjml runs
}
\author{
Marcos Alves
}
