% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMAPSPAM.R
\name{downloadMAPSPAM}
\alias{downloadMAPSPAM}
\title{downloadMAPSPAM}
\usage{
downloadMAPSPAM()
}
\value{
raw files for MAPSPAM
}
\description{
Downloads the MAP-SPAM (SPAM) data set for harvested and physical croparea
}
\examples{
\dontrun{
a <- download("downloadMAPSPAM")
}

}
\seealso{
[downloadSource()]
}
\author{
Edna J. Molina Bacca
}
