% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRamankutty.R
\name{downloadRamankutty}
\alias{downloadRamankutty}
\title{downloadRamankutty}
\usage{
downloadRamankutty()
}
\description{
download Ramankutty available land si (Source: Ramankutty N, Foley JA, Norman J and McSweeney K (2002) The global distribution of cultivable lands: current patterns and sensitivity to possible climate change. Global Ecology and Biogeography, 11, 377-392.)
}
\examples{
\dontrun{ a <- downloadSource() }

}
\seealso{
\code{\link{downloadSource}} \code{\link{readRamankutty}}
}
\author{
Felicitas Beier
}
