% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSchulteUebbing.R
\name{downloadSchulteUebbing}
\alias{downloadSchulteUebbing}
\title{downloadSchulteUebbing}
\usage{
downloadSchulteUebbing()
}
\value{
A list containing metadata from Schulte-Uebbing et al. 2022.
}
\description{
Download the critical nitrogen surplus values of Schulte-Uebbing et al. 2022.
  This function requires the 'archive' package, which is suggested. If 'archive' is not available,
  the function will stop with an error.
}
\examples{
\dontrun{
  downloadSchulteUebbing()
}
}
