% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmagpie-package.R
\docType{package}
\name{mrmagpie-package}
\alias{mrmagpie}
\alias{mrmagpie-package}
\title{mrmagpie: madrat based MAgPIE Input Data Library}
\description{
Provides functions for MAgPIE country and cellular input data generation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrmagpie}
  \item \doi{10.5281/zenodo.4319612}
  \item Report bugs at \url{https://github.com/pik-piam/mrmagpie/issues}
}

}
\author{
\strong{Maintainer}: Kristine Karstens \email{karstens@pik-potsdam.de}

Authors:
\itemize{
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}
  \item David Chen
  \item Michael Windisch
  \item Marcos Alves
  \item Felicitas Beier \email{beier@pik-potsdam.de}
  \item Alexandre Köberle \email{alexkob@pik-potsdam.de}
  \item Patrick v. Jeetze \email{vjeetze@pik-potsdam.de}
  \item Abhijeet Mishra \email{mishra@pik-potsdam.de}
  \item Florian Humpenoeder \email{humpenoeder@pik-potsdam.de}
  \item Pascal Sauer \email{pascal.sauer@pik-potsdam.de}
  \item Patrick Rein \email{patrick.rein@pik-potsdam.de}
}

}
