% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBendingTheCurve.R
\name{readBendingTheCurve}
\alias{readBendingTheCurve}
\title{readBendingTheCurve}
\usage{
readBendingTheCurve(subtype)
}
\arguments{
\item{subtype}{Data used in the Bending the Curve initiative.
Type "rr_layer" for the range-size rarity layer and "luh2_side_layers" for the LUH2 Side Layers.}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read bending the curve data
}
\examples{

\dontrun{
  readSource("BendingTheCurve", subtype="rr_layer", convert="onlycorrect")
}

}
\author{
Patrick v. Jeetze
}
