% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiess2024.R
\name{readBiess2024}
\alias{readBiess2024}
\title{readBiess2024}
\usage{
readBiess2024(subtype = "MRI-ESM2-0:ssp370:multi", subset = seq(1966, 2100, 1))
}
\arguments{
\item{subtype}{climatemodel:scenario:variable
- climate model: specific (e.g., "MRI-ESM2-0") or "all"
- scenario: ssp126, ssp245, ssp370 available
- variable: Occurrence of "p_e" (dry) , "pr" (wet), or "tasmax" (heat) extreams or multi
  (any of the three) (0 or 1) in one year}

\item{subset}{Years to read as a vector}
}
\value{
MAgPIE object of the defined data
}
\description{
Biess data contains information on 2.5° x 2.5° on occurence of extreme events based on different
             CMIP6 gcm data. This information based on the paper doi: 10.1088/1748-9326/ad4619.
}
\examples{
\dontrun{
readSource("Biess2024", subtype = "MRI-ESM2-0:ssp370:multi", subset = seq(1995, 2100, 1))
}
}
\author{
Edna Molina Bacca
}
