% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCO2Atmosphere_new.R
\name{readCO2Atmosphere_new}
\alias{readCO2Atmosphere_new}
\title{readCO2Atmosphere}
\usage{
readCO2Atmosphere_new(subtype = "ISIMIP3b:ssp126")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
Magpie objects with results on global level
}
\description{
Read CO2 global atmospheric concentration
}
\examples{

\dontrun{
readSource("CO2Atmosphere_new", subtype = "ISIMIP3b:ssp126", convert = FALSE)
}

}
\author{
Marcos Alves, Kristine Karstens
}
