% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDinerstein2017.R
\name{readDinerstein2017}
\alias{readDinerstein2017}
\title{readDinerstein2017}
\usage{
readDinerstein2017(subtype = "grassland")
}
\arguments{
\item{subtype}{Biome group whose areal cover fraction is returned. Currently only
\code{"grassland"} = RESOLVE biomes 7 (tropical and subtropical grasslands, savannas and
shrublands), 8 (temperate grasslands, savannas and shrublands), 9 (flooded grasslands and
savannas) and 10 (montane grasslands and shrublands).}
}
\value{
magpie object (67420 cells) with the cover fraction of the selected biome group.
Cells outside any ecoregion of that group are NA and set to zero in \code{\link{correctDinerstein2017}}.
}
\description{
Reads the RESOLVE Terrestrial Ecoregions 2017 (Dinerstein et al. 2017)
and returns the areal cover fraction of a biome group per 0.5 degree cell.
}
\examples{
\dontrun{
readSource("Dinerstein2017", subtype = "grassland", convert = "onlycorrect")
}

}
\seealso{
\code{\link{correctDinerstein2017}}
}
\author{
Florian Humpenoeder
}
