% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEvapotranspiration.R
\name{readEvapotranspiration}
\alias{readEvapotranspiration}
\title{readEvapotranspiration}
\usage{
readEvapotranspiration(subtype = "H08:mri-esm2-0:historical")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
MAgPIE objects with results on cellular level.
}
\description{
Read evapotranspiration data
}
\examples{
\dontrun{
readSource("Evapotranspiration", subtype, convert = "onlycorrect")
}

}
\seealso{
\code{\link{readEvapotranspiration}}
}
\author{
Marcos Alves
}
