% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFishcatches.R
\name{readFishCatches}
\alias{readFishCatches}
\title{readFishCatches}
\usage{
readFishCatches()
}
\value{
Magpie object with results on cellular level for soil types
}
\description{
Read soil classification data used as input for lpjml
}
\examples{

\dontrun{
readSource("SoilClassification")
}

}
\author{
Marcos Alves, Kristine Karstens
}
