% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGCMClimate_new.R
\name{readGCMClimate_new}
\alias{readGCMClimate_new}
\title{readGCMClimate_new}
\usage{
readGCMClimate_new(
  subtype = "ISIMIP3b:IPSL-CM6A-LR:historical:1850-2014:tas",
  subset = "annualmean"
)
}
\arguments{
\item{subtype}{Switch between different inputs,
e.g. "ISIMIP3b:IPSL-CM6A-LR:historical:1850-2014:tas"
Available variables are: * tas -
                         * wet -
                         * per -}

\item{subset}{Switch between different subsets of the same subtype}
}
\value{
MAgPIE objects with results on cellular level.
}
\description{
Read Climate data used as LPJmL inputs into MAgPIE objects
}
\examples{
\dontrun{
readSource("GCMClimate_new", subtype, convert = "onlycorrect")
}

}
\seealso{
\code{\link{readGCMClimate_new}}
}
\author{
Marcos Alves, Kristine Karstens, Felicitas Beier
}
