% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPD2022.R
\name{readGPD2022}
\alias{readGPD2022}
\title{readGPD2022}
\usage{
readGPD2022()
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
read x
Data from the Global Peatland Database provided by Alexandra Barthelmes.
The original xls file has been clean-up manually (country names). Turkey had
two identical entries in the original xls file.
Sources:
"Inventory Reports and National Communications UNFCC 2014",
"soil and peatland science",
"European Mires Book" ,
"own estimates (incl. GIS data)",
}
\examples{
\dontrun{
readSource("x", convert = "onlycorrect")
}
}
\author{
Florian Humpenoeder
}
