% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPM2.R
\name{readGPM2}
\alias{readGPM2}
\title{readGPM2}
\usage{
readGPM2(subtype = "1km")
}
\arguments{
\item{subtype}{resolution ("1km" or "500m")}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
read peatland area from GPM2
}
\examples{
\dontrun{
readSource("GPM2", convert = "onlycorrect")
}
}
\author{
Florian Humpenoeder
}
