% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGrassSoilEmu.R
\name{readGrassSoilEmu}
\alias{readGrassSoilEmu}
\title{readGrassSoilEmu}
\usage{
readGrassSoilEmu(
  subtype = "ISIMIP3b:IPSL_CM6A_LR:ssp126:1965_2100:5f5fa2:weights"
)
}
\arguments{
\item{subtype}{Subtype of file to be opened. Subtypes available:
'weights', 'inputs',  'stddevs' and 'means'.}
}
\value{
Magpie objects with a diverse inforamtion
}
\description{
Read files related to the training and optimization of the LPJml emulators.
}
\examples{
\dontrun{
readSource("GrassSoilEmu", subtype =
"ISIMIP3b:IPSL_CM6A_LR:ssp126:1965_2100:5f5fa2:weights", convert = F)
}

}
\author{
Marcos Alves
}
