% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGridPopGao.R
\name{readGridPopGao}
\alias{readGridPopGao}
\title{readGridPopGao}
\usage{
readGridPopGao()
}
\description{
Read gridded population, by urban and rural, from Gao O'Neill and JOnes dataset,
see https://www.cgd.ucar.edu/iam/modeling/spatial-population-scenarios.html https://doi.org/10.7927/m30p-j498
}
\author{
David Chen
}
