% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGridPopIsimip.R
\name{readGridPopIsimip}
\alias{readGridPopIsimip}
\title{readGridPopIsimip}
\usage{
readGridPopIsimip(subtype)
}
\arguments{
\item{subtype}{past (1965-2005) or future (2010-2100)}
}
\value{
A MAgPIE object, cellular 0.5deg resolution, of population (millions)
}
\description{
Reads in past and future (SSP1-5) gridded population data, from
ISIMIP database, Past data is based on HYDE3.2, while future SSPs are based
on projections from Jones & O'Neill 2016
}
\author{
David Chen, Marcos Alves, Felicitas Beier
}
