% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIrrigationSystem.R
\name{readIrrigationSystem}
\alias{readIrrigationSystem}
\title{readIrrigationSystem}
\usage{
readIrrigationSystem(subtype = "Jaegermeyr_magpiecell")
}
\arguments{
\item{subtype}{Data source to be read from: LPJmL for LPJmL-dominant-irrigation-system-type or Jaegermeyr for country-specific shares and lpjcell (67420) or magpiecell (59199)}
}
\value{
MAgPIE object of at country-level
}
\description{
Read in irrigation system type for initialization
}
\examples{
\dontrun{ readSource("IrrigationSystem", convert="onlycorrect") }

}
\author{
Felicitas Beier
}
