% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLabourProdImpactEmu.R
\name{readLabourProdImpactEmu}
\alias{readLabourProdImpactEmu}
\title{readLabourProdImpactEmu}
\usage{
readLabourProdImpactEmu()
}
\value{
magpie object of gridded productivity loss in percent (0-100)
}
\description{
read in labour productivity impacts from climate change emulated by the LAMACLIMA project

based on method of Orlov et al. 2019. Economics of Disasters and Climate Change, 3(3), 191-211.
}
\seealso{
\code{\link{readSource}}
}
\author{
Michael Windisch, Florian Humpenöder, Felicitas Beier
}
