% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMehta2024.R
\name{readMehta2024}
\alias{readMehta2024}
\title{readMehta2024}
\usage{
readMehta2024(subtype = "v4_GMIA")
}
\arguments{
\item{subtype}{data subtype to be downloaded
Subtypes available:
Combination of version (v3, v4) and data source:
'GMIA': gridded base map for downscaling from Stefan et al. (2013).
        Global Map of Irrigation Areas version 5.
'Meier2018': gridded base map for downscaling from Meier, et al. (2018).
             Global Irrigated Areas.
Separated by "_" (e.g., subtype = "v4_GMIA")}
}
\description{
reads in Global Area Equipped for Irrigation for years 1900-2015 from Mehta et al. (2022)
}
\examples{
\dontrun{
a <- readSource("Mehta2024")
}
}
\seealso{
[correctMehta2024()]
}
\author{
Felicitas Beier
}
