% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPastYields.R
\name{readPastYields}
\alias{readPastYields}
\title{readPastYields}
\usage{
readPastYields(subtype)
}
\arguments{
\item{subtype}{type of Pasture Management}
}
\description{
Function `readPastYields` read binary data from output of LPJmL runs and processes them into a `magpie object`.
}
\author{
Marcos Alves
}
