% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRamankutty.R
\name{readRamankutty}
\alias{readRamankutty}
\title{readRamankutty}
\usage{
readRamankutty()
}
\value{
magpie object
}
\description{
Read in data of Ramankutty dataset (Source: Ramankutty N, Foley JA, Norman J and McSweeney K (2002) The global distribution of cultivable lands: current patterns and sensitivity to possible climate change. Global Ecology and Biogeography, 11, 377-392.). Note: data processing analogous to /p/projects/magpie/data/input/other/rev23/avl_land_si.R
}
\examples{
\dontrun{
  readSource("Ramankutty", convert="onlycorrect")
}
}
\author{
Felicitas Beier
}
