% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSchulteUebbing.R
\name{readSchulteUebbing}
\alias{readSchulteUebbing}
\title{readSchulteUebbing}
\usage{
readSchulteUebbing()
}
\value{
a magpie object containing critical nitrogen surplus
}
\description{
read the critical nitrogen surplus dataset of Schulte-Uebbing et al. 2022
}
\examples{
\dontrun{
readSource("SchulteUebbing")
}
}
\author{
Michael Crawford
}
