% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTravelTimeNelson2019.R
\name{readTravelTimeNelson2019}
\alias{readTravelTimeNelson2019}
\title{readTravelTimeNelson2019}
\usage{
readTravelTimeNelson2019(subtype = "cities")
}
\arguments{
\item{subtype}{cities or ports data}
}
\value{
gridded magpie object for 2015, minimum travel time to cities in minutes
}
\description{
Read minimum travel time to cities and ports and
ports of various size, see metadata file in source folder
}
\author{
David M Chen
}
