% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTRENDY.R
\name{readTRENDY}
\alias{readTRENDY}
\title{readTRENDY}
\usage{
readTRENDY(subtype)
}
\arguments{
\item{subtype}{name of the TRENDY dataset to read. Current options include:
"CABLEPOP", "CLASSIC", "CLM5.0", "DLEM", "ED", "ELM",
"IBIS", "ISAM", "ISBACTRIP", "JSBACH", "JULES",
"LPJ-GUESS", "LPJml", "LPJwsl", "lpxqs", "OCN",
"ORCHIDEE", "SDGVM", "VISIT", "YIBS"}
}
\value{
a named list of terra::rast objects
}
\description{
read a TRENDY dataset
}
\examples{
\dontrun{
readSource("TRENDY", subtype = "JSBACH")
}
}
\author{
Michael Crawford
}
