% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUrbanLandGao.R
\name{readUrbanLandGao}
\alias{readUrbanLandGao}
\title{readUrbanLandGao}
\usage{
readUrbanLandGao()
}
\value{
magpie object of 2000-2100 urban land in Mha, in 10 year intervals
}
\description{
Read gridded urban land, from Gao O'Neill and JOnes SEDAC dataset,https://sedac.ciesin.columbia.edu/data/set/ssp-1-8th-urban-land-extent-projection-base-year-ssp-2000-2100
}
\author{
David M Chen
}
