% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGrassProdDisagg.R
\name{toolGrassProdDisagg}
\alias{toolGrassProdDisagg}
\title{Disaggregate grassland areas}
\usage{
toolGrassProdDisagg(grass_prod_lr,land_hr, weight_factor, map_file, lpjml_yields, ite)
}
\arguments{
\item{grass_prod_lr}{grassland production in low resolution to be disaggregated}

\item{land_hr}{Magpie object with celular cost for each pasture management type}

\item{weight_factor}{Magpie object with celular yields for each management type}

\item{map_file}{cluster mappint to cells and regions}

\item{lpjml_yields}{Lpjml yeilds used for magpie optimization}

\item{ite}{number of redistribution iterations}
}
\value{
Mapie object with the total pasture areas optimized for fulfulling pasture demand in the available areas. An extra dimention is added to
capture cells that are infeasible.
}
\description{
Disaggregate grassland areas and move values based on a maximum theorical regional productivity
}
\examples{
\dontrun{
toolGrassProdDisagg(grass_prod_lr,land_hr, weight_factor, map_file, lpjml_yields, ite)
}

}
\author{
Marcos Alves
}
