% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolNeuralNet.R
\name{toolNeuralNet}
\alias{toolNeuralNet}
\title{Neural Network Reconstruction}
\usage{
toolNeuralNet(inputsMl, weights, activation)
}
\arguments{
\item{inputsMl}{Neural Network input features properly scaled with
the scale and center attributes of the scaled training set in a matrix format.}

\item{weights}{The learned weights and biases in a list format as outputed by the
function \code{keras::get_weights()}.}

\item{activation}{Name of the activation function used for training.
Currently implemented functions: `relu`, `softplus`, `sigmoid`. Optionally, a custom
activation function can be passed using a "." to indicate where the layer inputs should be piped.}
}
\value{
The evaluated result of the neural network for the \code{input_ml} parameter.
}
\description{
Reconstructs and evaluate a neural network from the weights and biases provided as arguments
}
\author{
Marcos Alves
}
