% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRefoldWeights.R
\name{toolRefoldWeights}
\alias{toolRefoldWeights}
\title{Refold weights from NN training
Refold weights into their original configuration.}
\usage{
toolRefoldWeights(x)
}
\arguments{
\item{x}{magpie object containing weights.}
}
\description{
Refold weights from NN training
Refold weights into their original configuration.
}
\author{
Marcos Alves
}
