% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAgeClassDistribution.R
\name{calcAgeClassDistribution}
\alias{calcAgeClassDistribution}
\title{calcAgeClassDistribution}
\usage{
calcAgeClassDistribution(cells = "lpjcell")
}
\arguments{
\item{cells}{lpjcell for 67420 cells or magpiecell for 59199 cells}
}
\value{
magpie object in cluster resolution
}
\description{
This function calculates the share of each age class in secondary forests in each
             MAgPIE simulation cluster based on Global Forest Age Dataset from Poulter et al. 2019
}
\examples{
\dontrun{
calcOutput("AgeClassDistribution", aggregate = FALSE)
}

}
\author{
Abhijeet Mishra, Felicitas Beier
}
