% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvlCropland.R
\name{calcAvlCropland}
\alias{calcAvlCropland}
\title{calcAvlCropland}
\usage{
calcAvlCropland(cropland_scen = "all", cells = "magpiecell")
}
\arguments{
\item{cropland_scen}{different options are
\itemize{
\item \code{"all_marginal"}: Include all marginal land
\item \code{"half_marginal"}: Half of the marginal land is excluded
\item \code{"no_marginal"}: Marginal land is fully excluded
\item \code{"all"}: Returns all of the above options
}}

\item{cells}{magpiecell (59199 cells) or lpjcell (67420 cells)}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates the total available cropland per grid cell, based on physical
cropland suitability data or other criteria, such as constraints on cropland expansion
}
\examples{
\dontrun{
calcOutput("AvlCropland", cropland_scen = "all", cells = "magpiecell", aggregate = FALSE)
}

}
\author{
Patrick v. Jeetze
}
