% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvlCropland.R
\name{calcAvlCropland}
\alias{calcAvlCropland}
\title{calcAvlCropland}
\usage{
calcAvlCropland(cropland_scen = "all", cells = "magpiecell")
}
\arguments{
\item{cropland_scen}{different options are
\itemize{
\item \code{"allmarginal_0pNonCropVeg"}: All marginal land, 0 \% conservation of non-cropland vegetation
\item \code{"allmarginal_10pNonCropVeg"}: All marginal land, 10 \% conservation of non-cropland vegetation
\item \code{"allmarginal_20pNonCropVeg"}: All marginal land, 20 \% conservation of non-cropland vegetation
\item \code{"allmarginal_30pNonCropVeg"}: All marginal land, 30 \% conservation of non-cropland vegetation
\item \code{"halfmarginal_0pNonCropVeg"}: Half of the marginal land excluded, 0 \% conservation of non-cropland vegetation
\item \code{"halfmarginal_10pNonCropVeg"}: Half of the marginal land excluded, 10 \% conservation of non-cropland vegetation
\item \code{"halfmarginal_20pNonCropVeg"}: Half of the marginal land excluded, 20 \% conservation of non-cropland vegetation
\item \code{"halfmarginal_30pNonCropVeg"}: Half of the marginal land excluded, 30 \% conservation of non-cropland vegetation
\item \code{"nomarginal_0pNonCropVeg"}: No marginal land, 0 \% conservation of non-cropland vegetation
\item \code{"nomarginal_10pNonCropVeg"}: No marginal land, 10 \% conservation of non-cropland vegetation
\item \code{"nomarginal_20pNonCropVeg"}: No marginal land, 20 \% conservation of non-cropland vegetation
\item \code{"nomarginal_30pNonCropVeg"}: No marginal land, 30 \% conservation of non-cropland vegetation
\item \code{"all"}: Returns all information of the above options
}}

\item{cells}{magpiecell (59199 cells) or lpjcell (67420 cells)}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates the total available cropland per grid cell, based on physical
cropland suitability data or other criteria, such as constraints on cropland expansion
}
\examples{
\dontrun{
calcOutput("AvlCropland", cropland_scen = "all", cells = "magpiecell", aggregate = FALSE)
}

}
\author{
Patrick v. Jeetze
}
