% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvlCropland.R
\name{calcAvlCropland}
\alias{calcAvlCropland}
\title{calcAvlCropland}
\usage{
calcAvlCropland(cropland_scen = "all", cells = "magpiecell")
}
\arguments{
\item{cropland_scen}{different are
\itemize{
\item \code{"allmarginal_0pNonCropVeg"}:
\item \code{"allmarginal_10pNonCropVeg"}:
\item \code{"allmarginal_20pNonCropVeg"}:
\item \code{"allmarginal_30pNonCropVeg"}:
\item \code{"halfmarginal_0pNonCropVeg"}:
\item \code{"halfmarginal_10pNonCropVeg"}:
\item \code{"halfmarginal_20pNonCropVeg"}:
\item \code{"halfmarginal_30pNonCropVeg"}:
\item \code{"nomarginal_0pNonCropVeg"}:
\item \code{"nomarginal_10pNonCropVeg"}:
\item \code{"nomarginal_20pNonCropVeg"}:
\item \code{"nomarginal_30pNonCropVeg"}:
\item \code{"all"}:
}}

\item{cells}{magpiecell (59199 cells) or lpjcell (67420 cells)}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates the total available cropland per grid cell, based on physical
cropland suitability data or other criteria, such as constraints on cropland expansion
}
\examples{
\dontrun{
calcOutput("AvlCropland", cropland_scen = "all", cells = "magpiecell", aggregate = FALSE)
}

}
\author{
Patrick v. Jeetze
}
