% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvlLandSi.R
\name{calcAvlLandSi}
\alias{calcAvlLandSi}
\title{calcAvlLandSi}
\usage{
calcAvlLandSi(cells = "magpiecell")
}
\arguments{
\item{cells}{magpiecell (59199 cells) or lpjcell (67420 cells)}
}
\value{
magpie object in cellular resolution
}
\description{
Extracts si0 and nsi0 areas based on Ramankutty dataset
}
\examples{
\dontrun{
calcOutput("AvlLandSi", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
