% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvlWater.R
\name{calcAvlWater}
\alias{calcAvlWater}
\title{calcAvlWater}
\usage{
calcAvlWater(
  selectyears = "all",
  version = "LPJmL4",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  time = "raw",
  averaging_range = NULL,
  dof = NULL,
  harmonize_baseline = FALSE,
  ref_year = "y2015",
  EFR = TRUE
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{version}{Switch between LPJmL4 and LPJmL5}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{time}{Time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{harmonize_baseline}{FALSE (default): no harmonization, TRUE: if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{Reference year for harmonization baseline (just specify when harmonize_baseline=TRUE)}

\item{EFR}{Environmental flow requirements activated (TRUE) or not (FALSE)}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates water availability for MAgPIE retrieved from LPJmL using a river routing algorithm for distribution of discharge across the river basin
}
\examples{
\dontrun{ calcOutput("AvlWater", aggregate = FALSE) }

}
\author{
Jens Heinke, Felicitas Beier
}
