% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbon.R, R/calcCarbonTests.R
\name{calcCarbon}
\alias{calcCarbon}
\title{calcCarbon}
\usage{
calcCarbon(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_84a69edd", crop =
    "ggcmi_phase3_nchecks_72c185fa"),
  climatetype = "GSWP3-W5E5:historical",
  stage = "raw"
)

calcCarbon(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_84a69edd", crop =
    "ggcmi_phase3_nchecks_72c185fa"),
  climatetype = "GSWP3-W5E5:historical",
  stage = "raw"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different GCM climate scenarios}

\item{stage}{Switch for raw data or harmonization}
}
\value{
magpie object in cellular resolution

magpie object in cellular resolution
}
\description{
This function extracts carbon densities from LPJ to MAgPIE

This function extracts carbon densities from LPJ to MAgPIE
}
\examples{
\dontrun{ calcOutput("Carbon", aggregate = FALSE) }

\dontrun{ calcOutput("CarbonTests", aggregate = FALSE) }

}
\author{
Kristine Karstens

Kristine Karstens, Florian Humpenoeder
}
