% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbon.R
\name{calcCarbon}
\alias{calcCarbon}
\title{calcCarbon}
\usage{
calcCarbon(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_44ac93de", crop =
    "ggcmi_phase3_nchecks_9ca735cb"),
  climatetype = "GSWP3-W5E5:historical"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different GCM climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts carbon densities from LPJ to MAgPIE
}
\examples{
\dontrun{ calcOutput("Carbon", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
