% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbon.R
\name{calcCarbon}
\alias{calcCarbon}
\title{calcCarbon}
\usage{
calcCarbon(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_84a69edd", crop =
    "ggcmi_phase3_nchecks_72c185fa"),
  climatetype = "GSWP3-W5E5:historical",
  pastsc2natvegsc = FALSE
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different GCM climate scenarios}

\item{pastsc2natvegsc}{Test switch for substituting pasture soil carbon to nat vet soil carbon}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts carbon densities from LPJ to MAgPIE
}
\examples{
\dontrun{ calcOutput("Carbon", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
