% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbon_new.R
\name{calcCarbon_new}
\alias{calcCarbon_new}
\title{calcCarbon_new}
\usage{
calcCarbon_new(lpjml = c(natveg = "LPJmL4", crop = "LPJmL5"), climatetype)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different GCM climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts carbon densities from LPJ to MAgPIE
}
\examples{
\dontrun{ calcOutput("Carbon_new", aggregate = FALSE) }

}
\author{
Kristine Karstens
}
