% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCellCountryFraction.R
\name{calcCellCountryFraction}
\alias{calcCellCountryFraction}
\title{calcCellCountryFraction}
\usage{
calcCellCountryFraction(cells = "lpjcell")
}
\arguments{
\item{cells}{lpjcell for 67420 cells or magpiecell for 59199 cells}
}
\value{
Clustered MAgPIE object on requested resolution
}
\description{
cell fraction belonging to a country based on LanduseInitialisation
}
\examples{
\dontrun{
calcOutput("calcCellCountryFraction", aggregate = FALSE)
}

}
\author{
Florian Humpenoeder
}
