% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDegradationYieldReduction.R
\name{calcDegradationYieldReduction}
\alias{calcDegradationYieldReduction}
\title{calcDegradationYieldReduction}
\usage{
calcDegradationYieldReduction()
}
\value{
magpie object in cellular resolution
}
\description{
Function creates dummy file for including yield reduction coefficients to represent land degradation
}
\examples{
\dontrun{
calcOutput("DegradationYieldReduction", aggregate = FALSE)
}
}
\author{
Patrick v. Jeetze
}
