% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnvmtlFlow.R
\name{calcEnvmtlFlow}
\alias{calcEnvmtlFlow}
\title{calcEnvmtlFlow}
\usage{
calcEnvmtlFlow(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_84a69edd", crop =
    "ggcmi_phase3_nchecks_72c185fa"),
  climatetype = "GSWP3-W5E5:historical",
  stage = "harmonized2020",
  LFR_val = 0.1,
  HFR_LFR_less10 = 0.2,
  HFR_LFR_10_20 = 0.15,
  HFR_LFR_20_30 = 0.07,
  HFR_LFR_more30 = 0,
  seasonality = "grper"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{stage}{Degree of processing: raw, smoothed, harmonized, harmonized2020}

\item{LFR_val}{Strictness of environmental flow requirements}

\item{HFR_LFR_less10}{High flow requirements (share of total water for cells) with LFR<10percent of total water}

\item{HFR_LFR_10_20}{High flow requirements (share of total water for cells) with 10percent < LFR < 20percent of total water}

\item{HFR_LFR_20_30}{High flow requirements (share of total water for cells) with 20percent < LFR < 30percent of total water}

\item{HFR_LFR_more30}{High flow requirements (share of total water for cells) with LFR>30percent of total water}

\item{seasonality}{grper (default): EFR in growing period per year; total: EFR throughout the year; monthly: monthly EFRs}
}
\value{
magpie object in cellular resolution
}
\description{
This function calculates environmental flow requirements (EFR) for MAgPIE retrieved from LPJmL monthly discharge and water availability
}
\examples{
\dontrun{ calcOutput("EnvmtlFlow", aggregate = FALSE) }

}
\author{
Felicitas Beier, Abhijeet Mishra
}
