% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEvapotranspiration.R
\name{calcEvapotranspiration}
\alias{calcEvapotranspiration}
\title{calcEvapotranspiration}
\usage{
calcEvapotranspiration(subtype = "H08:mri-esm2-0")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
magpie object in cellular resolution
}
\description{
Calc evapotranspiration data for SSP cenarios in mm/month
}
\examples{
\dontrun{
calcOutput("Evapotranspiration", subtype = "H08:mri-esm2-0")
}

}
\author{
Marcos Alves
}
