% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcExtremesFrequency.R
\name{calcExtremesFrequency}
\alias{calcExtremesFrequency}
\title{calcExtremesFrequency}
\usage{
calcExtremesFrequency(
  subtype = "MRI-ESM2-0:ssp370:multi",
  initialYear = 1995,
  yearsOver = 30,
  gcmAvg = FALSE,
  percentage = FALSE,
  method = "Biess2024"
)
}
\arguments{
\item{subtype}{details of climate model, scenario, and extreme variable in the format: gcm:scenario:variable}

\item{initialYear}{Initial year of the magpie object, e.g., 1995}

\item{yearsOver}{Occurence over xx years. E.g., Occurence over 20 years}

\item{gcmAvg}{when all climate models are read, should the funcion average overall of them}

\item{percentage}{should the output be given as a percentage (TRUE) or years over xx years (FALSE) or
as fraction ("fraction")}

\item{method}{the source of the extremes data}
}
\value{
magpie object
}
\description{
It calculates the number of extreams events occurring in a define number of years.
E.g., Occurence over 20 years.
}
\examples{
\dontrun{
calcOutput("ExtremesFrequency", subtype = "MRI-ESM2-0:ssp370:multi", initialYear = 1995)
}

}
\author{
Edna J. Molina Bacca
}
