% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGCMClimate.R
\name{calcGCMClimate}
\alias{calcGCMClimate}
\title{calcGCMClimate}
\usage{
calcGCMClimate(subtype = "HadGEM2_ES:rcp8p5.temperature")
}
\arguments{
\item{subtype}{type of climate data to collect}
}
\value{
magpie object in cellular resolution
}
\description{
Disaggregate CO2 global atmospheric concentration to cellular level
}
\examples{
\dontrun{ calcOutput("GCMClimate", subtype = "HadGEM2_ES:rcp8p5.temperature") }

}
\author{
Marcos Alves, Kristine Karstens
}
