% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrasslandsYields.R
\name{calcGrasslandsYields}
\alias{calcGrasslandsYields}
\title{calcGrasslandsYields}
\usage{
calcGrasslandsYields(
  lpjml = "lpjml5p2_pasture",
  climatetype = "MRI-ESM2-0:ssp370",
  cells = "lpjcell",
  subtype = "/co2/Nreturn0p5",
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  past_mngmt = "mdef"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Global Circulation Model to be used}

\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}

\item{subtype}{Switch between different climate scenarios}

\item{lsu_levels}{Livestock unit levels in the source folder}

\item{past_mngmt}{pasture areas management option}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates rangelands maximum output and managed pastures yields
}
\examples{
\dontrun{
calcOutput("GrasslandsYields")
}
}
\author{
Marcos Alves
}
