% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrasslandsYields.R
\name{calcGrasslandsYields}
\alias{calcGrasslandsYields}
\title{calcGrasslandsYields}
\usage{
calcGrasslandsYields(
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  past_mngmt = "me2",
  subtype = "lpjml5p2_pasture:IPSL_CM6A_LR:ssp126_co2_limN"
)
}
\arguments{
\item{lsu_levels}{Livestock unit levels in the source folder}

\item{past_mngmt}{pasture areas management option}

\item{subtype}{Switch between different climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates rangelands maximum output and managed pastures yields
}
\examples{
\dontrun{
calcOutput("GrasslandsYields", lsu_levels, past_mngmt = "2me", subtype)
}

}
\author{
Marcos Alves
}
