% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrasslandsYields.R
\name{calcGrasslandsYields}
\alias{calcGrasslandsYields}
\title{calcGrasslandsYields}
\usage{
calcGrasslandsYields(
  lpjml = lpjml[["grass"]],
  climatetype = climatetype,
  subtype = "/co2/Nreturn0p5/limN",
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  past_mngmt = "me2"
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Global Circulation Model to be used}

\item{subtype}{Switch between different climate scenarios}

\item{lsu_levels}{Livestock unit levels in the source folder}

\item{past_mngmt}{pasture areas management option}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates rangelands maximum output and managed pastures yields
}
\examples{
\dontrun{
calcOutput("GrasslandsYields", lsu_levels, past_mngmt = "2me", subtype)
}

}
\author{
Marcos Alves
}
