% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIrrigation.R
\name{calcIrrigation}
\alias{calcIrrigation}
\title{calcIrrigation}
\usage{
calcIrrigation(
  lpjml = c(natveg = "LPJmL4_for_MAgPIE_44ac93de", crop =
    "ggcmi_phase3_nchecks_9ca735cb"),
  climatetype = "GSWP3-W5E5:historical",
  rainfedweight = 0.01
)
}
\arguments{
\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{rainfedweight}{For clustering airrig is weighted with
cropland_irrigated + rainfedweight * cropland_rainfed (default: 0.01)}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts irrigation water (airrig: water applied additionally to rainfall)
             from LPJmL to MAgPIE
}
\examples{
\dontrun{
calcOutput("Irrigation", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Abhijeet Mishra
}
