% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLsuDensityHist.R
\name{calcLsuDensityHist}
\alias{calcLsuDensityHist}
\title{calcLsuDensityHist}
\usage{
calcLsuDensityHist(disagg_type = "grassland")
}
\arguments{
\item{disagg_type}{select the disaggregaton weights for biomass production (can be either grassland or livestock)}
}
\value{
List of magpie object with results on cluster level
}
\description{
Calculate livestock historical livestock densities
}
\examples{

\dontrun{
calcOutput("LsuDensityHist")
}
}
\author{
Marcos Alves
}
