% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMowing.R
\name{calcMowing}
\alias{calcMowing}
\title{calcMowing}
\usage{
calcMowing(
  mowing_events = "2me",
  lpjml = "LPJml_mowing",
  climatetype = "HadGEM2_ES:rcp8p5:co2"
)
}
\arguments{
\item{mowing_events}{number of mowing events per year expressed as `2me`}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates mowing pasture yields
}
\examples{
\dontrun{
calcOutput("Mowing", mowing_events = "2me", lpjml = "LPJml_mowing", climatetype)
}

}
\author{
Marcos Alves
}
