% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastrTauHist.R
\name{calcPastrTauHist}
\alias{calcPastrTauHist}
\title{calcPastrTauHist}
\usage{
calcPastrTauHist()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates managed pastures Tau based on FAO yield trends for 1995.
}
\examples{
\dontrun{
calcOutput("PastrTauHist")
}

}
\author{
Marcos Alves
}
