% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastrTauHist.R
\name{calcPastrTauHist}
\alias{calcPastrTauHist}
\title{calcPastrTauHist}
\usage{
calcPastrTauHist(past_mngmt = "mdef", cells = "lpjcell")
}
\arguments{
\item{past_mngmt}{Pasture management reference yield}

\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates managed pastures Tau based on FAO yield trends for 1995.
}
\examples{
\dontrun{
calcOutput("PastrTauHist", past_mngmt)
}

}
\author{
Marcos Alves
}
