% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastureSuit.R
\name{calcPastureSuit}
\alias{calcPastureSuit}
\title{calcPastureSuit}
\usage{
calcPastureSuit(
  datasource = "ISIMIP3bv2",
  climatetype = "MRI-ESM2-0:ssp126",
  lpjml = "LPJmL4_for_MAgPIE_44ac93de",
  smoothPrecipitation = 10,
  smoothOut = 10
)
}
\arguments{
\item{datasource}{Document}

\item{climatetype}{Document}

\item{lpjml}{Document}

\item{smoothPrecipitation}{Smooth precipitation climate data over time}

\item{smoothOut}{smooth the Pasture suitability areas variations over time}
}
\value{
List of magpie object with results on cluster level
}
\description{
Calculate glassland suitable for pasture management based
on population and aridity criteria.
}
\examples{
\dontrun{
calcOutput("PastureSuit")
}
}
\author{
Marcos Alves
}
