% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastureSuit.R
\name{calcPastureSuit}
\alias{calcPastureSuit}
\title{calcPastureSuit}
\usage{
calcPastureSuit(subtype = "ISIMIP3b:IPSL-CM6A-LR:1850-2100")
}
\arguments{
\item{subtype}{Select version, climate model and period.}
}
\value{
List of magpie object with results on cluster level
}
\description{
Calculate glassland suitable for pasture management based on population and aridity criteria.
}
\examples{

\dontrun{
calcOutput("PastureSuit")
}
}
\author{
Marcos Alves
}
