% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastureSuit.R
\name{calcPastureSuit}
\alias{calcPastureSuit}
\title{calcPastureSuit}
\usage{
calcPastureSuit(
  subtype = "ISIMIP3bv2:IPSL-CM6A-LR:1850-2100",
  smoothPrecipitation = 10,
  smoothOut = 10
)
}
\arguments{
\item{subtype}{Select version, climate model and period.}

\item{smoothPrecipitation}{Smooth precipitation climate data over time}

\item{smoothOut}{smooth the Pasture suitability areas variations over time}
}
\value{
List of magpie object with results on cluster level
}
\description{
Calculate glassland suitable for pasture management based on population and aridity criteria.
}
\examples{
\dontrun{
calcOutput("PastureSuit")
}
}
\author{
Marcos Alves
}
