% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPastYields.R
\name{calcPastYields}
\alias{calcPastYields}
\title{calcPastYields}
\usage{
calcPastYields(
  lsu_levels = c(seq(0, 2, 0.2), 2.5),
  mowing_events = "2me",
  lpjml = c(cgrazing = "LPJmL_cgrazing", mowing = "LPJmL_mowing"),
  climatetype = "HadGEM2_ES:rcp8p5:co2"
)
}
\arguments{
\item{lsu_levels}{Livestock unit levels in the source folder}

\item{mowing_events}{number of mowing events per year expressed as `2me`}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates continuous grazing maximum output and mowing yields
}
\examples{
\dontrun{calcOutput("PastYields", lsu_levels, mowing_events = "2me", lpjml, climatetype)}

}
\author{
Marcos Alves
}
