% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRiskYieldExtremes.R
\name{calcRiskYieldExtremes}
\alias{calcRiskYieldExtremes}
\title{calcRiskYieldExtremes}
\usage{
calcRiskYieldExtremes(
  subtype = "MRI-ESM2-0:ssp370",
  initialYear = 2025,
  endYear = 2100,
  yearsOver = 30,
  kcr = "all",
  version = "ggcmi_phase3_nchecks_bft_6277d36e",
  timeStep = 5,
  extremeType = "multi"
)
}
\arguments{
\item{subtype}{details of climate model and scenario}

\item{initialYear}{Initial year of the magpie object. Due to data availability first year available is 2010}

\item{endYear}{End year of the magpie object, e.g., 2100}

\item{yearsOver}{Window of years for the moving average. E.g., 30}

\item{kcr}{"all" or specific crop type to calculate vulnerability}

\item{version}{"version of LPJmL to use}

\item{timeStep}{years difference between time steps of the output}

\item{extremeType}{Type of extreme "p_e" (dry) , "pr" (wet), or "tasmax" (heat) extreams or multi
(any of the three)}
}
\value{
magpie object
}
\description{
It calculates the anomalies from a [certain number of years] moving average for LPJmL data,
calculated as (x-xAvg)/xAvg
}
\examples{
\dontrun{
calcOutput("calcRiskYieldExtremes", subtype = "MRI-ESM2-0:ssp370", initialYear = 1995,
                                 yearsOver = 20, kcr = "all")
}

}
\author{
Edna J. Molina Bacca
}
