% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcScaleEnvironmentData.R
\name{calcScaleEnvironmentData}
\alias{calcScaleEnvironmentData}
\title{calcScaleEnvironmentData}
\usage{
calcScaleEnvironmentData(
  climatetype = "HadGEM2_ES:rcp8p5:co2",
  sar = 20,
  sel_feat = c("lsu_ha", "temperature", "precipitation", "longwave_radiation",
    "shortwave_radiation", "wetdays", "CO2ATMconcentration", "Ks", "Sf", "w_pwp", "w_fc",
    "w_sat", "hsg", "soilc")
)
}
\arguments{
\item{climatetype}{Switch between different climate scenarios}

\item{sar}{Average range for smoothing annual variations}

\item{sel_feat}{features names to be included in the output file}
}
\value{
magpie object in cellular resolution
}
\description{
Scale climate, CO2 and soil environmental conditions on cellular level
}
\examples{
\dontrun{
calcOutput("ScaleEnvironmentData", climatetype = "HadGEM2_ES:rcp8p5:co2", sar = 20, sel_feat)
}

}
\author{
Marcos Alves
}
