% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransportCosts_new.R
\name{calcTransportCosts_new}
\alias{calcTransportCosts_new}
\title{calcTransportCosts_new}
\usage{
calcTransportCosts_new(transport = "all")
}
\arguments{
\item{transport}{"all" or "nonlocal". "all" means all production incurs transport costs,
while "nonlocal" sees only production greater than local rural consumption with transport costs}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates country-level transport costs from GTAP total
transport costs, cellular production, and cellular travel time
}
\examples{
\dontrun{
calcOutput("TransportCosts_new")
}
}
\seealso{
[calcTransportTime()],
[calcGTAPTotalTransportCosts()]
}
\author{
David M Chen
}
