% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransportTime.R
\name{calcTransportTime}
\alias{calcTransportTime}
\title{calcTransportTime}
\usage{
calcTransportTime(subtype = "cities50", cells = "lpjcell")
}
\arguments{
\item{subtype}{currently only cities of 5, 20, or 50 thousand people ("cities5", "cities20", "cities50") or
ports of various sizes ("portsLarge|Medium|Small|VerySmall|Any")}

\item{cells}{number of cells to be returned: magpiecell (59199), lpjcell (67420)}
}
\value{
magpie object in cellular resolution
}
\description{
Function extracts travel time to major cities in minutes
}
\examples{
\dontrun{
calcOutput("TransportTime", aggregate = FALSE)
}

}
\author{
David Chen
}
