% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterUseNonAg.R
\name{calcWaterUseNonAg}
\alias{calcWaterUseNonAg}
\title{calcWaterUseNonAg}
\usage{
calcWaterUseNonAg(
  selectyears = "all",
  source = "WATCH_ISIMIP_WATERGAP",
  finalcells = "magpiecell",
  time = "raw",
  averaging_range = NULL,
  dof = NULL,
  lpjml = c(natveg = "LPJmL4", crop = "LPJmL5"),
  seasonality = "grper",
  climatetype = "HadGEM2_ES:rcp2p6:co2",
  harmonize_baseline = "CRU_4",
  ref_year = "y2015"
)
}
\arguments{
\item{selectyears}{years to be returned}

\item{source}{data source to be used (e.g. WATERGAP2020)}

\item{finalcells}{number of cells to be returned by this function magpiecell (59199) or lpjcell (67420)}

\item{time}{Time smoothing: average, spline or raw (default)}

\item{averaging_range}{only specify if time=="average": number of time steps to average}

\item{dof}{only specify if time=="spline": degrees of freedom needed for spline}

\item{lpjml}{Defines LPJmL version for crop/grass and natveg specific inputs}

\item{seasonality}{grper (default): non-agricultural water demand in growing period per year; total: non-agricultural water demand throughout the year}

\item{climatetype}{switch between different climate scenarios (default: "CRU_4") for calcGrowingPeriod}

\item{harmonize_baseline}{FALSE (default), if a baseline is specified here input data is harmonized to that baseline (from ref_year onwards) Note: only applies for calcGrowingPeriod}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year for calcGrowingPeriod}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts non-agricultural water demand
}
\examples{
\dontrun{ calcOutput("WaterUseNonAg", aggregate=FALSE) }

}
\author{
Felicitas Beier
}
