% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYieldsMAgPIE.R
\name{calcYieldsMAgPIE}
\alias{calcYieldsMAgPIE}
\title{calcYieldsMAgPIE}
\usage{
calcYieldsMAgPIE(
  datasource = list(lpjml = "ggcmi_phase3_nchecks_9ca735cb", isimip = NULL),
  climatetype = "GSWP3-W5E5:historical",
  selectyears = seq(1965, 2100, by = 5),
  multicropping = FALSE,
  calibration = NULL,
  weighting = "totalCrop",
  marginal_land = "q33_marginal:rainfed_and_irrigated",
  indiaYields = FALSE,
  scaleFactor = 0.3
)
}
\arguments{
\item{datasource}{Choose LPJmL version for main crop inputs and optionally ISIMIP version
For ISIMIP choose crop model/gcm/rcp/co2 combination formatted like this:
"yields:EPIC-IIASA:ukesm1-0-ll:ssp585:default:3b"}

\item{climatetype}{Climate scenario or historical baseline}

\item{selectyears}{Years to be returned}

\item{multicropping}{Multicropping activated (TRUE) or not (FALSE) and
Multiple Cropping Suitability mask selected}

\item{calibration}{NULL for uncalibrated yields or list for FAO calibration}

\item{weighting}{area-based weighting options for yield aggregation}

\item{marginal_land}{marginal land setting used for weighting}

\item{indiaYields}{If TRUE returns scaled yields for rainfed crops in India}

\item{scaleFactor}{Integer value by which indiaYields will be scaled}
}
\value{
magpie object in cellular resolution
}
\description{
Orchestrates MAgPIE yields by combining LPJmL-based yields, optional ISIMIP
replacement, India yield scaling, and aggregation weights.
}
\examples{
\dontrun{
calcOutput("YieldsMAgPIE", aggregate = FALSE)
}

}
\author{
Kristine Karstens, Felicitas Beier
}
