% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEvapotranspiration.R
\name{readEvapotranspiration}
\alias{readEvapotranspiration}
\title{readEvapotranspiration}
\usage{
readEvapotranspiration(subtype = "H08:mri-esm2-0:historical")
}
\arguments{
\item{subtype}{Switch between different inputs eg. "ISIMIP3b:IPSL-CM6A-LR:historical:1850-2014:tas"}
}
\value{
MAgPIE objects with results on cellular level.
}
\description{
Read Climate data used as LPJmL inputs into MAgPIE objects
}
\examples{
\dontrun{
readSource("Evapotranspiration", subtype, convert = "onlycorrect")
}

}
\seealso{
\code{\link{readEvapotranspiration}}
}
\author{
Marcos Alves, Kristine Karstens
}
