% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMAPSPAM.R
\name{readMAPSPAM}
\alias{readMAPSPAM}
\title{readMAPSPAM}
\usage{
readMAPSPAM(type = "harvested", year = 2000)
}
\arguments{
\item{type}{It can be either "harvested" or "physical" area}

\item{year}{year of data (2000, 2005, 2010)}
}
\value{
magpie object with croparea data in ha
}
\description{
Reads the MAP-SPAM crop data per year (mapping each year different)
}
\examples{
\dontrun{
a <- readSource("MAPSPAM")
}

}
\seealso{
[readSource()]
}
\author{
Edna J. Molina Bacca
}
