% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMAPSPAM.R
\name{readMAPSPAM}
\alias{readMAPSPAM}
\title{readMAPSPAM}
\usage{
readMAPSPAM(subtype = "harvested")
}
\arguments{
\item{subtype}{It can be either "harvested" or "physical" area}
}
\value{
magpie object with croparea data in ha
}
\description{
Reads the MAP-SPAM crop data per year (mapping each year different)
}
\examples{
\dontrun{
a <- readSource("MAPSPAM")
}
}
\seealso{
[readSource()]
}
\author{
Edna J. Molina Bacca, Felicitas Beier
}
