% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRamankutty.R
\name{readRamankutty}
\alias{readRamankutty}
\title{readRamankutty}
\usage{
readRamankutty()
}
\value{
magpie object
}
\description{
Read in data of Ramankutty dataset
}
\examples{
\dontrun{
  readSource("Ramankutty", convert="onlycorrect")
}

}
\author{
Felicitas Beier
}
