% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRamankutty.R
\name{readRamankutty}
\alias{readRamankutty}
\title{readRamankutty}
\usage{
readRamankutty()
}
\value{
magpie object
}
\description{
Read in data of Ramankutty dataset
(Source: Ramankutty N, Foley JA, Norman J and McSweeney K (2002) The global distribution of cultivable lands: current patterns and sensitivity to possible climate change. Global Ecology and Biogeography, 11, 377-392.).
Link to data: https://www.nelson.wisc.edu/sage/data-and-models/global-land-use/grid.php
}
\examples{
\dontrun{
  readSource("Ramankutty", convert="onlycorrect")
}

}
\author{
Felicitas Beier
}
