% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTravelTimeNelson2019.R
\name{readTravelTimeNelson2019}
\alias{readTravelTimeNelson2019}
\title{readTravelTimeNelson2019}
\usage{
readTravelTimeNelson2019(subtype = "cities50")
}
\arguments{
\item{subtype}{currently only cities of 5, 20, or 50 thousand people ("cities5", "cities20", "cities50") or
ports of various sizes ("portsLarge|Medium|Small|VerySmall|Any")}
}
\value{
gridded magpie object for 2015, minimum travel time to cities in minutes
}
\description{
Read minimum travel time to cities and ports and
ports of various size, see metadata file in source folder
}
\author{
David M Chen
}
