% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTRENDY.R
\name{readTRENDY}
\alias{readTRENDY}
\title{readTRENDY}
\usage{
readTRENDY(subtype)
}
\arguments{
\item{subtype}{name of the TRENDY dataset to read. Current options includes:
"CABLEPOP", "CARDAMOM", "CLASSIC", "CLM5.0", "DLEM",
"ELM", "IBIS", "ISBACTRIP", "JSBACH", "LPJ-GUESS",
"LPJml", "LPJwsl", "lpxqs", "OCN", "ORCHIDEE",
"SDGVM", "VISIT"}
}
\value{
a named list of terra::rast objects
}
\description{
read a TRENDY dataset
}
\examples{
\dontrun{
readSource("TRENDY", subtype = "JSBACH")
}
}
\author{
Michael Crawford
}
